% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{all_concat}
\alias{all_concat}
\title{all_concat}
\usage{
all_concat(..., sep = "_")
}
\arguments{
\item{...}{is all the vectors of the elements that will be concatenated, see examples}

\item{sep}{is the separator to use between concatenated elements}
}
\description{
Output all the possible concatenations of elements in different vectors, see examples
}
\examples{

print(all_concat(c("France", "Germany"), c("2012", "2013"), c(1:2), sep = "_"))

[1] "France_2012_1"  "Germany_2012_1" "France_2013_1"  "Germany_2013_1"
[5] "France_2012_2"  "Germany_2012_2" "France_2013_2"  "Germany_2013_2"

}
