% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{all_stat}
\alias{all_stat}
\title{all_stat}
\usage{
all_stat(inpt_v, var_add = c(), stat_var = c(), inpt_df)
}
\arguments{
\item{inpt_v}{is the modalities of the variables}

\item{var_add}{is the variables you want to get the stats from}

\item{stat_var}{is the stats indicators you want}

\item{inpt_df}{is the input dataframe}
}
\description{
Allow to see all the main statistics indicators (mean, median, variance, standard deviation, sum, max, min, quantile) of variables in a dataframe by the modality of a variable in a column of the input datarame. In addition to that, you can get the occurence of other qualitative variables by your chosen qualitative variable, you have just to precise it in the vector "stat_var" where all the statistics indicators are given with "occu-var_you_want/".
}
\examples{
df <- data.frame("mod"=c("first", "seco", "seco", "first", "first", "third", "first"), 
                "var1"=c(11, 22, 21, 22, 22, 11, 9), 
               "var2"=c("d", "d", "z", "z", "z", "d", "z"), 
               "var3"=c(45, 44, 43, 46, 45, 45, 42),
              "var4"=c("A", "A", "A", "A", "B", "C", "C"))

all_stat(inpt_v=c("first", "seco"), var_add = c("var1", "var2", "var3", "var4"), 
 stat_var=c("sum", "mean", "median", "sd", "occu-var2/", "occu-var4/", "variance", "quantile-0.75/"), 
 inpt_df=df)
}
