% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{and_bool2}
\alias{and_bool2}
\title{and_bool2}
\usage{
and_bool2(inpt_datf)
}
\description{
Returns a boolean vector according to 'and' condition on boolean vector in a dataframe, each column represents a boolean vector, see examples, uses a different algorithm than and_bool1
}
\examples{

print(and_bool2(inpt_datf = data.frame("bool1" = c(TRUE, FALSE, FALSE, TRUE),
                                       "bool2" = c(TRUE, TRUE, FALSE, TRUE))))

[1]  TRUE FALSE FALSE  TRUE


}
