% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{append_row}
\alias{append_row}
\title{append_row}
\usage{
append_row(df_in, df, hmn = 1, na_col = c(), unique_do_not_know = NA)

append_row(df_in, df, hmn = 1, na_col = c(), unique_do_not_know = NA)
}
\arguments{
\item{df_in}{is the dataframe from which the row will append to another or the same  dataframe}

\item{df}{is the dataframe to which the row will append}

\item{hmn}{is how many time the last row will be appended}

\item{na_col}{is a vector containing the columns that won't append and will be replaced by another value (unique_do_not_know)}

\item{unique_do_not_know}{is the value of the non appending column in the appending row}
}
\description{
Append the last row from dataframe to the another or same dataframe

Append the last row from dataframe to the another or same dataframe
}
