% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{appndr}
\alias{appndr}
\title{appndr}
\usage{
appndr(inpt_v, val = NA, hmn, strt = "max")
}
\arguments{
\item{inpt_v}{is the input vector}

\item{val}{is the special value}

\item{hmn}{is the number of special value element added}

\item{strt}{is the index from which appending begins, defaults to max which means the end of "inpt_v"}
}
\description{
Append to a vector "inpt_v" a special value "val" n times "mmn". The appending begins at "strt" index.
}
\examples{

print(appndr(inpt_v=c(1:3), val="oui", hmn=5))

#[1] "1"   "2"   "3"   "oui" "oui" "oui" "oui" "oui"

print(appndr(inpt_v=c(1:3), val="oui", hmn=5, strt=1))

#[1] "1"   "oui" "oui" "oui" "oui" "oui" "2"   "3" 

}
