% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{arroundr_mean}
\alias{arroundr_mean}
\title{arroundr_mean}
\usage{
arroundr_mean(inpt_v = c(), step_val)
}
\arguments{
\item{inpt_v}{is the input vector}

\item{step_val}{is the step_value}
}
\description{
Takes an ascendly int ordered vector as input and assigns each elements that are close enough to the same value accrdng to a step value (step_value), see examples.
}
\examples{

x <- arroundr_mean(inpt_v = c(-11:25), step_val = 5)
print(x)
print(length(x))

[1] -9.0 -9.0 -9.0 -9.0 -9.0 -4.0 -4.0 -4.0 -4.0 -4.0  1.0  1.0  1.0  1.0  1.0
[16]  6.0  6.0  6.0  6.0  6.0 11.0 11.0 11.0 11.0 11.0 16.0 16.0 16.0 16.0 16.0
[31] 21.0 21.0 21.0 21.0 21.0 23.8 23.8
[1] 37

}
