% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{arroundr_min}
\alias{arroundr_min}
\title{arroundr_min}
\usage{
arroundr_min(inpt_v = c(), step_val)
}
\arguments{
\item{inpt_v}{is the input vector}

\item{step_val}{is the step value}
}
\description{
Takes an ascendly int ordered vector as input and assigns each elements that are close enough to the same value accrdng to a step value (step_value), see examples.
}
\examples{

print(arroundr_min(inpt_v = c(-11:25), step_val = 5))

 [1] -11 -11 -11 -11 -11 -11  -6  -6  -6  -6  -6  -1  -1  -1  -1  -1   4   4   4
[20]   4   4   9   9   9   9   9  14  14  14  14  14  19  19  19  19  19  24

}
