% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{better_match}
\alias{better_match}
\title{better_match}
\usage{
better_match(inpt_v = c(), ptrn, untl = 1, nvr_here = NA)
}
\arguments{
\item{inpt_v}{is the input vector}

\item{ptrn}{is the pattern to be matched}

\item{untl}{is the maximum number of matched pattern outputed}

\item{nvr_here}{is a value you are sure is not present in inpt_v}
}
\description{
Allow to get the nth element matched in a vector
}
\examples{

print(better_match(inpt_v=c(1:12, 3, 4, 33, 3), ptrn=3, untl=1))

#[1] 3

print(better_match(inpt_v=c(1:12, 3, 4, 33, 3), ptrn=3, untl=5))
 
#[1]  3 13 16

print(better_match(inpt_v=c(1:12, 3, 4, 33, 3), ptrn=c(3, 4), untl=5))
 
[1]  3 13 16  4 14

print(better_match(inpt_v=c(1:12, 3, 4, 33, 3), ptrn=c(3, 4), untl=c(1, 5)))

[1]  3  4 14

}
