% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{better_split}
\alias{better_split}
\title{better_split}
\usage{
better_split(inpt, split_v = c())
}
\arguments{
\item{inpt}{is the input character}

\item{split_v}{is the vector containing the splits}
}
\description{
Allows to split a string by multiple split, returns a vector and not a list.
}
\examples{

print(better_split(inpt = "o-u_i", split_v = c("-")))

[1] "o"   "u_i"

print(better_split(inpt = "o-u_i", split_v = c("-", "_")))

[1] "o" "u" "i"

}
