% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{better_split_any}
\alias{better_split_any}
\title{better_split_any}
\usage{
better_split_any(inpt, split_v = c())
}
\arguments{
\item{inpt}{is the input character}

\item{split_v}{is the vector containing the splits}
}
\description{
Allows to split a string by multiple split regardless of their length, returns a vector and not a list. Contrary to better_split, this functions keep the delimiters in the output.
}
\examples{

print(better_split_any(inpt = "o-u_i", split_v = c("-")))

[1] "o"  "-" "u_i"

print(better_split_any(inpt = "o-u_i", split_v = c("-", "_")))

[1] "o" "-" "u"  "_" "i"

print(better_split_any(inpt = "--o--_/m/m/__-opo-/m/-u_i-_--", split_v = c("--", "_", "/")))

 [1] "--"    "o"     "--"    "_"     "/"     "m"     "/"     "m"     "/"    
[10] "_"     "_"     "-opo-" "/"     "m"     "/"     "-u"    "_"     "i-"   
[19] "_"     "--"   


print(better_split_any(inpt = "(ok(ee:56))(ok2(oui)(ee:4))", split_v = c("(", ")", ":")))

 [1] "("   "ok"  "("   "ee"  ":"   "56"  ")"   ")"   "("   "ok2" "("   "oui"
 [13] ")"   "("   "ee"  ":"   "4"   ")"   ")"  

}
