% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{better_sub}
\alias{better_sub}
\title{better_sub}
\usage{
better_sub(inpt_v = c(), pattern, replacement, untl_v = c())
}
\arguments{
\item{inpt_v}{is a vector containing all the elements that contains expressions to be substituted}

\item{pattern}{is the expression that will be substituted}

\item{replacement}{is the expression that will substituate pattern}

\item{untl_v}{is a vector containing, for each element of inpt_v, the number of pattern that will be substituted}
}
\description{
Allow to perform a sub operation to a given number of matched patterns, see examples
}
\examples{

print(better_sub(inpt_v = c("yes NAME, i will call NAME and NAME", 
                            "yes NAME, i will call NAME and NAME"),
                 pattern = "NAME",
                 replacement = "Kevin",
                 untl = c(2)))

[1] "yes Kevin, i will call Kevin and NAME"
[2] "yes Kevin, i will call Kevin and NAME"

print(better_sub(inpt_v = c("yes NAME, i will call NAME and NAME", 
                            "yes NAME, i will call NAME and NAME"),
                 pattern = "NAME",
                 replacement = "Kevin",
                 untl = c(2, 3)))

[1] "yes Kevin, i will call Kevin and NAME" 
[2] "yes Kevin, i will call Kevin and Kevin"

print(better_sub(inpt_v = c("yes NAME, i will call NAME and NAME", 
                             "yes NAME, i will call NAME and NAME"),
                  pattern = "NAME",
                  replacement = "Kevin",
                  untl = c("max", 3)))

[1] "yes Kevin, i will call Kevin and Kevin"
[2] "yes Kevin, i will call Kevin and Kevin"

}
