% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{better_sub_mult}
\alias{better_sub_mult}
\title{better_sub_mult}
\usage{
better_sub_mult(
  inpt_v = c(),
  pattern_v = c(),
  replacement_v = c(),
  untl_v = c()
)
}
\arguments{
\item{inpt_v}{is a vector containing all the elements that contains expressions to be substituted}

\item{pattern_v}{is a vector containing all the patterns to be substituted in any elements of inpt_v}

\item{replacement_v}{is a vector containing the expression that are going to substituate those provided by pattern_v}

\item{untl_v}{is a vector containing, for each element of inpt_v, the number of pattern that will be substituted}
}
\description{
Allow to perform a sub_mult operation to a given number of matched patterns, see examples
}
\examples{

print(better_sub_mult(inpt_v = c("yes NAME, i will call NAME and NAME2", 
                             "yes NAME, i will call NAME and NAME2, especially NAME2"),
                  pattern_v = c("NAME", "NAME2"),
                  replacement_v = c("Kevin", "Paul"),
                  untl = c(1, 3)))

[1] "yes Kevin, i will call NAME and Paul"                 
[2] "yes Kevin, i will call NAME and Paul, especially Paul"

print(better_sub_mult(inpt_v = c("yes NAME, i will call NAME and NAME2", 
                              "yes NAME, i will call NAME and NAME2, especially NAME2"),
                   pattern_v = c("NAME", "NAME2"),
                   replacement_v = c("Kevin", "Paul"),
                   untl = c("max", 3)))

[1] "yes Kevin, i will call Kevin and Kevin2"                   
[2] "yes Kevin, i will call Kevin and Kevin2, especially Kevin2"

}
