% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{bind_cols}
\alias{bind_cols}
\title{bind_cols}
\usage{
bind_cols(from_datf, in_datf)
}
\arguments{
\item{from_datf}{is the dataframe that contains the cols to find among other cols}

\item{in_datf}{is the dataframe that only contans the cols to find in from_datf}
}
\description{
Allow to find the cols of a dataframe in an other dataframe, see examples
}
\examples{

iris[, 5] <- as.character(iris[, 5])
iris <- cbind(iris, iris[, 4])
from_datf <- iris
in_datf <- iris[, c(1, 2, 2, 2, 4)]
bind_cols(from_datf = from_datf,
          in_datf = in_datf)

[[1]]
[1] 1

[[2]]
[1] 2

[[3]]
[1] 2

[[4]]
[1] 2

[[5]]
[1] 4 6

}
