% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{bind_rows}
\alias{bind_rows}
\title{bind_rows}
\usage{
bind_rows(from_datf, in_datf)
}
\arguments{
\item{from_datf}{is the dataframe that contains the rows to find among other rows}

\item{in_datf}{is the dataframe that only contans the rows to find in from_datf}
}
\description{
Allow to find the rows of a dataframe in an other dataframe, see examples
}
\examples{

iris[, 5] <- as.character(iris[, 5])
from_datf <- iris
in_datf <- iris[c(4, 2, 23, 21, 11), ]

bind_rows(from_datf = from_datf,
          in_datf = in_datf)

[[1]]
[1] 4

[[2]]
[1] 2

[[3]]
[1] 23

[[4]]
[1] 21

[[5]]
[1] 11

}
