% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{calc_occu_v}
\alias{calc_occu_v}
\title{calc_occu_v}
\usage{
calc_occu_v(f_v, w_v, nvr_here = NA)
}
\description{
Rearanges the index of a vector "w_v" to match the occurences of the common elements in another vector "f_v"
}
\examples{
print(calc_occu_v(f_v=c("e", "a", "z", NA, "a"), w_v=c("a", "a", "z")))

[1] 1 3 2
}
