% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{change_date}
\alias{change_date}
\title{change_date}
\usage{
change_date(
  date_,
  sep_,
  day_ = NA,
  month_ = NA,
  year_ = NA,
  hour_ = NA,
  min_ = NA,
  second_ = NA,
  frmt = "snhdmy"
)

change_date(
  date_,
  sep_,
  day_ = NA,
  month_ = NA,
  year_ = NA,
  hour_ = NA,
  min_ = NA,
  second_ = NA,
  frmt = "snhdmy"
)
}
\arguments{
\item{date_}{is the input date}

\item{sep_}{is the date separator}

\item{day_}{is the day to add (can be negative)}

\item{month_}{is the month to add (can be negative)}

\item{year_}{is the year to add (can be negative)}

\item{hour_}{is the hour to add (can be negative)}

\item{min_}{is the minute to add (can be negative)}

\item{second_}{is the second to add (can be negative)}

\item{frmt}{is the format of the input date, (deault set to "snhdmy" (second, minute, hour, day, month, year), so all variable are taken in count),
if you only want to work with standard date for example change this variable to "dmy"}
}
\description{
Allow to add to a date second-minute-hour-day-month-year

Allow to add to a date second-minute-hour-day-month-year
}
