% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{chr_removr}
\alias{chr_removr}
\title{chr_removr}
\usage{
chr_removr(inpt_v, ptrn_v)
}
\arguments{
\item{inpt_v}{is the input vector containing all the elements that may have the characters to be removed}

\item{ptrn_v}{is the vector containing all the characters that will be removed}
}
\description{
Allow to remove certain characters contained in a vector "ptrn_v" from elements in a another vector "inpt_v".
}
\examples{
print(chr_removr(inpt_v=c("oui?", "!oui??", "non", "!non"), ptrn_v=c("?")))
[1] "oui"  "!oui" "non"  "!non"

print(chr_removr(inpt_v=c("oui?", "!oui??", "non", "!non"), ptrn_v=c("?", "!")))
[1] "oui" "oui" "non" "non"
}
