% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{closest_date}
\alias{closest_date}
\title{closest_date}
\usage{
closest_date(
  vec,
  date_,
  frmt,
  sep_ = "/",
  sep_vec = "/",
  only_ = "both",
  head = NA
)
}
\arguments{
\item{vec}{is a vector containing the dates to be compared to the input date}

\item{date_}{is the input date}

\item{frmt}{is the format of the input date, (deault set to "snhdmy" (second, minute, hour, day, month, year), so all variable are taken in count),
if you only want to work with standard date for example change this variable to "dmy"}

\item{sep_}{is the separator for the input date}

\item{sep_vec}{is the separator for the dates contained in vec}

\item{only_}{is can be changed to "+" or "-" to repectively only return the higher dates and the lower dates (default set to "both")}

\item{head}{is the number of dates that will be returned (default set to NA so all dates in vec will be returned)}
}
\description{
return the closest dates from a vector compared to the input date
}
