% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{clusterizer_v}
\alias{clusterizer_v}
\title{clusterizer_v}
\usage{
clusterizer_v(inpt_v, w_v = NA, c_val)
}
\arguments{
\item{inpt_v}{is the vector containing the sequence of number}

\item{w_v}{is the vector containing the elements related to inpt_v, defaults to NA}

\item{c_val}{is the accuracy of the clusterization}
}
\description{
Allow to output clusters of elements. Takes as input a vector "inpt_v" containing a sequence of number. Can also take another vector "w_v" that has the same size of inpt_v because its elements are related to it. The way the clusters are made is related to an accuracy value which is "c_val". It means that if the difference between the values associated to 2 elements is superior to c_val, these two elements are in distinct clusters.
}
\examples{
 print(clusterizer_v(inpt_v=sample.int(20, 26, replace=T), w_v=NA, c_val=0.9))

[[1]]
[[1]][[1]]
[1] "j" "v"

[[1]][[2]]
[1] "x"

[[1]][[3]]
[1] "e" "m" "p" "s" "t" "b" "q" "z" "f"

[[1]][[4]]
[1] "a" "i"

[[1]][[5]]
[1] "c" "n" "o" "g" "u" "y" "h" "l"

[[1]][[6]]
[1] "d" "r" "w" "k"


[[2]]
[1] "1"  "2"  "-"  "4"  "4"  "-"  "6"  "10" "-"  "12" "12" "-"  "14" "16" "-" 
[16] "18" "19"

print(clusterizer_v(inpt_v=sample.int(40, 26, replace=T), w_v=letters, c_val=0.29))


[[1]]
[[1]][[1]]
[1] "a" "b" "c" "d" "e" "f" "g" "h"

[[1]][[2]]
[1] "i" "j" "k" "l"

[[1]][[3]]
[1] "m" "n"

[[1]][[4]]
[1] "o" "p" "q" "r" "s" "t" "u" "v" "w" "x" "y" "z"


[[2]]
[1] "1"  "5"  "-"  "8"  "10" "-"  "12" "13" "-"  "15" "20"
}
