% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{col_to_row}
\alias{col_to_row}
\title{col_to_row}
\usage{
col_to_row(inpt_datf)
}
\arguments{
\item{inpt_datf}{is the inout dataframe}
}
\description{
Allow to reverse a dataframe (cols become rows and rows become cols)
}
\examples{

datf_test <- data.frame(c(1:11), c(11:1))

print(col_to_row(inpt_datf = datf_test))

  X1 X2 X3 X4 X5 X6 X7 X8 X9 X10 X11
1  1  2  3  4  5  6  7  8  9  10  11
2 11 10  9  8  7  6  5  4  3   2   1

}
