% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{converter_date}
\alias{converter_date}
\title{converter_date}
\usage{
converter_date(inpt_date, convert_to, frmt = "snhdmy", sep_ = "-")
}
\arguments{
\item{inpt_date}{is the input date}

\item{convert_to}{is the time unit the input date will be converted ("s", "n", "h", "d", "m", "y")}

\item{frmt}{is the format of the input date}

\item{sep_}{is the separator of the input date. For example this input date "12-07-2012" has "-" as a separator}
}
\description{
Allow to convert any date like second/minute/hour/day/month/year to either second, minute...year. The input date should not necessarily have all its time units (second, minute...) but all the time units according to a format. Example: "snhdmy" is for second, hour, minute, day, month, year. And "mdy" is for month, day, year.
}
\examples{

print(converter_date(inpt_date="14-04-11-2024", sep_="-", frmt="hdmy", convert_to="m"))

#[1] 24299.15

print(converter_date(inpt_date="14-04-11-2024", sep_="-", frmt="hdmy", convert_to="y"))

#[1] 2024.929

print(converter_date(inpt_date="14-04-11-2024", sep_="-", frmt="hdmy", convert_to="s"))

#[1] 63900626400

print(converter_date(inpt_date="63900626400", sep_="-", frmt="s", convert_to="y"))

#[1] 2024.929

print(converter_date(inpt_date="2024", sep_="-", frmt="y", convert_to="s"))

#[1] 63873964800

}
