% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{converter_format}
\alias{converter_format}
\title{converter_format}
\usage{
converter_format(inpt_val, sep_ = "-", inpt_frmt, frmt, default_val = "00")
}
\arguments{
\item{inpt_val}{is the input value that is linked to the format}

\item{sep_}{is the separator of the value in inpt_val}

\item{inpt_frmt}{is the format of the input value}

\item{frmt}{is the format you want to convert to}

\item{default_val}{is the default value given to the units that are not present in the input format}
}
\description{
Allow to convert a format to another
}
\examples{

print(converter_format(inpt_val="23-12-05-1567", sep_="-", 
                       inpt_frmt="shmy", frmt="snhdmy", default_val="00"))

#[1] "23-00-12-00-05-1567"

print(converter_format(inpt_val="23-12-05-1567", sep_="-", 
                       inpt_frmt="shmy", frmt="Pnhdmy", default_val="00"))

#[1] "00-00-12-00-05-1567"

}
