% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{cumulated_rows}
\alias{cumulated_rows}
\title{cumulated_rows}
\usage{
cumulated_rows(inpt_datf, values_v = c())
}
\arguments{
\item{inpt_datf}{is the input data.frame}

\item{values_v}{is a vector containing all the values that a cell has to equal to return a TRUE value in the output vector at the index corresponding to the row of the cell}
}
\description{
Output a vector of size that equals to the rows number of the input dataframe, with TRUE value at the indices corresponding to the row where at least a cell of any column is equal to one of the values inputed in \code{values_v}
}
\examples{

datf_teste <- data.frame(c(1:10), c(10:1))

print(datf_teste)

   c.1.10. c.10.1.
1        1      10
2        2       9
3        3       8
4        4       7
5        5       6
6        6       5
7        7       4
8        8       3
9        9       2
10      10       1

print(cumulated_rows(inpt_datf = datf_teste, values_v = c(2, 3)))

[1]   FALSE TRUE TRUE   FALSE   FALSE   FALSE   FALSE TRUE TRUE   FALSE

}
