% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{cumulated_rows_na}
\alias{cumulated_rows_na}
\title{cumulated_rows_na}
\usage{
cumulated_rows_na(inpt_datf)
}
\arguments{
\item{inpt_datf}{is the input data.frame}
}
\description{
Output a vector of size that equals to the rows number of the input dataframe, with TRUE value at the indices corresponding to the row where at least a cell of any column is equal to NA.
}
\examples{

datf_teste <- data.frame(c(1, 2, 3, 4, 5, NA, 7), c(10, 9, 8, NA, 7, 6, NA))

print(datf_teste)

  c.1..2..3..4..5..NA..7. c.10..9..8..NA..7..6..NA.
1                       1                        10
2                       2                         9
3                       3                         8
4                       4                        NA
5                       5                         7
6                      NA                         6
7                       7                        NA

print(cumulated_rows_na(inpt_datf = datf_teste))

[1] FALSE FALSE FALSE  TRUE FALSE  TRUE  TRUE

}
