% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{cutr_v}
\alias{cutr_v}
\title{cutr_v}
\usage{
cutr_v(inpt_v, untl = "min")
}
\arguments{
\item{inpt_v}{is the input vector}

\item{untl}{is the maximum size of nchar authorized by an element, defaults to "min", it means the shortest element in the list}
}
\description{
Allow to reduce all the elements in a vector to a defined size of nchar
}
\examples{

test_v <- c("oui", "nonon", "ez", "aa", "a", "dsfsdsds")

print(cutr_v(inpt_v=test_v, untl="min"))

#[1] "o" "n" "e" "a" "a" "d"

print(cutr_v(inpt_v=test_v, untl=3))

#[1] "oui" "non" "ez"  "aa"  "a"   "dsf"

}
