% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{data_gen}
\alias{data_gen}
\title{data_gen}
\usage{
data_gen(
  type_ = c("number", "mixed", "string"),
  strt_l = c(0, 0, 10),
  nb_r = c(50, 10, 40),
  output = "gened.csv",
  properties = c("1-5", "1-5", "1-5"),
  type_distri = c("random", "random", "random"),
  str_source = c("a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n",
    "o", "p", "q", "r", "s", "t", "u", "w", "x", "y", "z"),
  round_l = c(0, 0, 0),
  sep_ = ","
)
}
\arguments{
\item{type_}{is a vector for wich argument is a column, a column can be made of numbers ("number"), string ("string") or both ("mixed")}

\item{strt_l}{is a vector containing for each column the row from which the data will begin to be generated}

\item{nb_r}{is a vector containing for each column, the number of row full from generated data}

\item{output}{is the name of the output csv file}

\item{properties}{is linked to type_distri because it is the parameters ("min_val-max_val") for "random type", ("u-x") for the poisson distribution, ("u-d") for gaussian distribution}

\item{type_distri}{is a vector which, for each column, associate a type of distribution ("random", "poisson", "gaussian"), it meas that non only the number but also the length of the string will be randomly generated according to these distribution laws}

\item{str_source}{is the source (vector) from which the character creating random string are (default set to the occidental alphabet)}

\item{round_l}{is a vector which, for each column containing number, associate a round value}

\item{sep_}{is the separator used to write data in the csv}
}
\value{
new generated data in addition to saving it in the output
}
\description{
Allo to generate in a csv all kind of data you can imagine according to what you provide
}
