% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{data_meshup}
\alias{data_meshup}
\title{data_meshup}
\usage{
data_meshup(
  data,
  cols = NA,
  file_ = NA,
  sep_ = ";",
  organisation = c(2, 1, 0),
  unic_sep1 = "_",
  unic_sep2 = "-"
)

data_meshup(
  data,
  cols = NA,
  file_ = NA,
  sep_ = ";",
  organisation = c(2, 1, 0),
  unic_sep1 = "_",
  unic_sep2 = "-"
)
}
\arguments{
\item{data}{is the data provided (vector) each column is separated by a unic separator and each dataset from the same column is separated by another unic separator (ex: c("\emph{", c("d", "-", "e", "-", "f"), "}", c("a", "a1", "-", "b", "-", "c", "c1")"_")}

\item{cols}{is the colnames of the data generated in a csv}

\item{file_}{is the file to which the data will be outputed}

\item{sep_}{is the separator of the csv outputed}

\item{organisation}{is the way variables include themselves, for instance ,resuming precedent example, if organisation=c(1, 0) so the data output will be:
d, a
d, a1
e, c
f, c
f, c1}

\item{unic_sep1}{is the unic separator between variables (default is "_")}

\item{unic_sep2}{is the unic separator between datasets (default is "-")}
}
\description{
Allow to automatically arrange 1 dimensional data according to vector and parameters

Allow to automatically arrange 1 dimensional data according to vector and parameters
}
