% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{date_sort}
\alias{date_sort}
\title{date_sort}
\usage{
date_sort(vec, asc = F, sep = "-")
}
\arguments{
\item{vec}{is the vector containing the dates.}

\item{asc}{is a boolean variable, that if set to TRUE will sort the dates ascendely and descendely if set to FALSE}

\item{sep}{is the separator of the date strings ex: "11-12-1998" the separator is "-"}
}
\description{
Allow to ascendely or desendely sort dates in a vector.
}
