% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{datf_appendr}
\alias{datf_appendr}
\title{datf_appendr}
\usage{
datf_appendr(inpt_datf)
}
\arguments{
\item{inpt_datf}{is the input dataframe}
}
\description{
Allow to append all columns of a dataframe in a vector.
}
\examples{

datf_teste <- data.frame("col1" = c(1:5), "col2" = c(5:1))

print(datf_appendr(inpt_datf = datf_teste))

[1] 1 2 3 4 5 5 4 3 2 1

}
