% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{datf_insertr}
\alias{datf_insertr}
\title{datf_insertr}
\usage{
datf_insertr(inpt_datf, ids_vec, val_l)
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{ids_vec}{is the ids where the rows has to be inserted after}

\item{val_l}{is a list containing all the rows (vector) to be inserted, linked to eevery index within ids_vec}
}
\description{
Insert rows after certain indexes, see examples
}
\examples{

datf <- data.frame(c(1:4), c(4:1))
print(datf)

  c.1.4. c.4.1.
1      1      4
2      2      3
3      3      2
4      4      1

print(datf_insertr(inpt_datf = datf, ids_vec = c(1, 3), val_l = list(c("non", "non"), c("oui", "oui"))))

  c.1.4. c.4.1.
1       1      4
2     non    non
21      2      3
3       3      2
5     oui    oui
4       4      1

print(datf_insertr(inpt_datf = datf, ids_vec = c(1, 3), val_l = list(c("non", "non"))))

  c.1.4. c.4.1.
1       1      4
2     non    non
21      2      3
3       3      2
5     non    non
4       4      1

}
