% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{datf_row_appendr}
\alias{datf_row_appendr}
\title{datf_row_appendr}
\usage{
datf_row_appendr(inpt_datf)
}
\arguments{
\item{inpt_datf}{is the input dataframe}
}
\description{
Allow to append all rows of a dataframe in a vector.
}
\examples{

datf_teste <- data.frame("col1" = c(1:5), "col2" = c(5:1))

print(datf_appendr(inpt_datf = datf_teste))

col1 col2 col1 col2 col1 col2 col1 col2 col1 col2 
   1    5    2    4    3    3    4    2    5    1 

}
