% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{datf_row_appendr2}
\alias{datf_row_appendr2}
\title{datf_row_appendr2}
\usage{
datf_row_appendr2(inpt_datf, chs_type = "integer")
}
\arguments{
\item{inpt_datf}{is the inout dataframe}
}
\description{
Allow to append all rows of a dataframe in a vector,
specifying the column types ("integer" or "character"), see examples
}
\examples{

datf_teste <- data.frame("col1" = c(1:5), "col2" = c(5:1), 
  "col3" = c("oui", "oui", "oui", "non", "non"))

print(datf_row_appendr2(inpt_datf = datf_teste, chs_type = "integer"))

NULL

print(datf_row_appendr2(inpt_datf = datf_teste, chs_type = "character"))

 col1  col2  col3  col1  col2  col3  col1  col2  col3  col1  col2  col3  col1 
  "1"   "5" "oui"   "2"   "4" "oui"   "3"   "3" "oui"   "4"   "2" "non"   "5" 
 col2  col3 
  "1" "non" 

}
