% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{days_from_month}
\alias{days_from_month}
\title{days_from_month}
\usage{
days_from_month(date_, sep_)

days_from_month(date_, sep_)
}
\arguments{
\item{date_}{is the input date}

\item{sep_}{is the separator of the input date}
}
\description{
Allow to find the number of days month from a month date, take in count leap year

Allow to find the number of days month from a month date, take in count leap year
}
