% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{delta_normal}
\alias{delta_normal}
\title{delta_normal}
\usage{
delta_normal(results_v = c(), mean_inpt, sd_inpt)
}
\arguments{
\item{results_v}{is the input vector}

\item{mean_inpt}{is the mea of the normal distribution}

\item{sd_inpt}{is the standard deviation of the normal distribution}
}
\description{
Returns the cumulative difrerence between a dataset of values and the normal distribution density formula, see examples
}
\examples{

x <- rnorm(n = 10000, mean = 15, sd = 1)
print(delta_normal(results_v = x,
                        mean_inpt = 15,
                        sd_inpt = 3))

[1] 15904.85

x <- rnorm(n = 10000, mean = 15, sd = 2)
print(delta_normal(results_v = x,
                        mean_inpt = 15,
                        sd_inpt = 3))

[1] 8050.662

x <- rnorm(n = 10000, mean = 15, sd = 5)
print(delta_normal(results_v = x,
                        mean_inpt = 15,
                        sd_inpt = 3))

[1] 16219.87

x <- rnorm(n = 10000, mean = 15, sd = 4)
print(delta_normal(results_v = x,
                        mean_inpt = 15,
                        sd_inpt = 3))

[1] 8081.28

x <- rnorm(n = 10000, mean = 15, sd = 3)
print(delta_normal(results_v = x,
                        mean_inpt = 15,
                        sd_inpt = 3))

[1] 474.8151

}
