% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{delta_unif}
\alias{delta_unif}
\title{delta_unif}
\usage{
delta_unif(inpt_v, min_inpt, max_inpt)
}
\arguments{
\item{inpt_v}{is the input vector that may represent a uniform distribution}

\item{min_inpt}{is the minimum of your uniform distribution}

\item{max_inpt}{is the maximum of your uniform distribution}
}
\description{
Returns the cumulative difference between a known uniform distribution and the input vector that may be a uniform distribution of given min and max, see examples.
}
\examples{

print(delta_unif(inpt_v = runif(n = 5000, min = 12, max = 17), min_inpt = 12, max_inpt = 17))

[1] 170.5542

print(delta_unif(inpt_v = runif(n = 5000, min = 12, max = 17), min_inpt = 122, max_inpt = 177))

[1] 675102.8

}
