% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{df_tuned}
\alias{df_tuned}
\title{df_tuned}
\usage{
df_tuned(df, val_to_stop, index_rc = NA, included = "yes")
}
\arguments{
\item{df}{is the input data.frame}

\item{val_to_stop}{is the vector containing the values to stop}

\item{index_rc}{is the value for the jump (default set to NA so default will be first case)}

\item{included}{is if the values to stop has to be also returned in the vectors (defaultn set to "yes")}
}
\description{
Allow to return a list from a dataframe following these rules:
First situation, I want the vectors from the returned list be composed of values that are separated by special values contained in a vector
ex: data.frame(c(1, 1, 2, 1), c(1, 1, 2, 1), c(1, 1, 1, 2)) will return list(c(1, 1), c(1, 1, 1), c(1, 1, 1, 1)) or
list(c(1, 1, 2), c(1, 1, 1, 2), c(1, 1, 1, 1, 2)) if i have chosen to take in count the 2. As you noticed here the value to stop is 2 but it can be several contained in a vector
Second situation: I want to return a list for every jump of 3.
If i take this dataframe data.frame(c(1, 1, 2, 1, 4, 4), c(1, 1, 2, 1, 3, 3), c(1, 1, 1, 2, 3, 3)) it will return list(c(1, 1, 2), c(1, 4, 4), c(1, 1, 2), c(1, 3, 3), c(1, 1, 1), c(2, 3, 3))
}
