% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{diff_datf}
\alias{diff_datf}
\title{diff_datf}
\usage{
diff_datf(datf1, datf2)
}
\arguments{
\item{datf1}{is an an input dataframe}

\item{datf2}{is an an input dataframe}
}
\description{
Returns a vector with the coordinates of the cell that are not equal between 2 dataframes (row, column).
}
\examples{

datf1 <- data.frame(c(1:6), c("oui", "oui", "oui", "oui", "oui", "oui"), c(6:1))

datf2 <- data.frame(c(1:7), c("oui", "oui", "oui", "oui", "non", "oui", "zz"))

print(diff_datf(datf1=datf1, datf2=datf2)) 

#[1] 5 1 5 2

}
