% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{diff_vec}
\alias{diff_vec}
\title{diff_vec}
\usage{
diff_vec(vec1, vec2)
}
\arguments{
\item{vec1}{is one of the two vectors}

\item{vec2}{is one of the two vectors}
}
\description{
Compare two vecters and return a boolean vector with TRUE if same element at the same index and FALSE if not
}
