% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{diff_xlsx}
\alias{diff_xlsx}
\title{diff_xlsx}
\usage{
diff_xlsx(
  file_,
  sht,
  v_old_begin,
  v_old_end,
  v_new_begin,
  v_new_end,
  df2 = NA,
  overwrite = T,
  color_ = "red",
  pattern = "",
  output = "out.xlsx",
  new_val = T,
  pattern_only = T
)
}
\arguments{
\item{file_}{is the file where the data is}

\item{sht}{is the sheet where the data is}

\item{v_old_begin}{is the corrdinate (row, column) where the data to be compared starts}

\item{v_old_end}{is the same but for its end}

\item{v_new_begin}{is the coordinates where the comparator data starts}

\item{v_new_end}{is the same but for its end}

\item{df2}{is optional, if the comparator dataset is directly a dataframe}

\item{overwrite}{allow to overwrite differences is (set to T by default)}

\item{color_}{is the color the differences will be outputed}

\item{pattern}{is the pattern that will be added to the differences if overwritten is set to TRUE}

\item{output}{is the name of the outputed xlsx (can be set to NA if no output)}

\item{new_val}{if overwrite is TRUE, then the differences will be overwritten by the comparator data}

\item{pattern_only}{will cover differences by pattern if overwritten is set to TRUE}
}
\description{
Allow to see the difference between two datasets and output it into an xlsx file
}
