% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{edm1_random_val}
\alias{edm1_random_val}
\title{edm1_random_val}
\usage{
edm1_random_val(len_untl = 5)
}
\arguments{
\item{len_untl}{is the maximum length that the random number could have}
}
\description{
Allow to generate a random number until a maximum length, see examples
}
\examples{

print(edm1_random_val(len_untl = 5))

[1] 54656

print(edm1_random_val(len_untl = 8))

[1] 64021015

print(edm1_random_val(len_untl = 3))

[1] 45

print(edm1_random_val(len_untl = 4))

[1] 6146

print(edm1_random_val(len_untl = 1))

[1] 3

}
