% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{edm1_random_val_spe}
\alias{edm1_random_val_spe}
\title{edm1_random_val_spe}
\usage{
edm1_random_val_spe(len_inpt = 5)
}
\arguments{
\item{len_inpt}{i the lenght of the random number that will be generated}
}
\description{
Allow to generate a random number of a pre determined length, see examples.
}
\examples{

print(edm1_random_val_spe(len_inpt = 5))

[1] 55272

print(edm1_random_val_spe(len_inpt = 8))

[1] 79930782

print(edm1_random_val_spe(len_inpt = 3))

[1] 480

print(edm1_random_val_spe(len_inpt = 4))

[1] 6865

print(edm1_random_val_spe(len_inpt = 1))

[1] 2

}
