% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{edm1_runif_deterministic}
\alias{edm1_runif_deterministic}
\title{edm1_runif_deterministic}
\usage{
edm1_runif_deterministic(n_inpt, min_inpt, max_inpt)
}
\arguments{
\item{n_inpt}{is the number of wanted values}
}
\description{
Produces a deterministic uniform distribution, see examples
}
\examples{

x <- edm1_runif_deterministic(n_inpt = 5000, min_inpt = 10, max_inpt = 15)

sd(x)

[1] 1.44352

sd(runif(n = 5000, min = 10, max = 15))

[1] 1.449532

x <- edm1_runif_deterministic(n_inpt = 5000, min_inpt = 10, max_inpt = 115)

sd(x)

[1] 30.31392

sd(runif(n = 5000, min = 10, max = 115))

[1] 30.33717

}
