% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{edm_group_by2}
\alias{edm_group_by2}
\title{edm_group_by2}
\usage{
edm_group_by2(inpt_datf, grp_v)
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{grp_v}{is the vector containiong the column names or the column numbers to perform the group by, see examples}
}
\description{
Performs a group by (different algorythm that edm_group_by1), see examples
}
\examples{

datf <- data.frame("col1" = c("A", "B", "B", "A", "C", "B"), 
                  "col2" = c("E", "R", "E", "E", "R", "R"), 
                  "col3" = c("P", "P", "O", "O", "P", "O"))
print(datf)

  col1 col2 col3
1    A    E    P
2    B    R    P
3    B    E    O
4    A    E    O
5    C    R    P
6    B    R    O

print(edm_group_by2(inpt_datf = datf, grp_v = c("col1")))

  col1 col2 col3
1    A    E    P
4    A    E    O
2    B    R    P
3    B    E    O
6    B    R    O
5    C    R    P

print(edm_group_by2(inpt_datf = datf, grp_v = c("col1", "col2")))

  col1 col2 col3
1    A    E    P
4    A    E    O
3    B    E    O
2    B    R    P
6    B    R    O
5    C    R    P

print(edm_group_by2(inpt_datf = datf, grp_v = c("col2", "col1")))

  col2 col1 col3
1    E    A    P
4    E    A    O
3    E    B    O
2    R    B    P
6    R    B    O
5    R    C    P

}
