% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{edm_pivot_series}
\alias{edm_pivot_series}
\title{edm_pivot_series}
\usage{
edm_pivot_series(inpt_datf, time_col, col_v = NULL)
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{time_col}{is the column name or number of the datafame}

\item{col_v}{is a vector containing all the column numbers or names of the variables, if null all the column will be considered as variables apart from the column designated in time_col}
}
\description{
Allow to create a new column for the value of the chosen columns at each new value of the column that represents the time. The occurence of each time stamp has to be equal, see examples (if not consider performing the time_serie_equalizer function fromm the same package)
}
\examples{

print(cur_datf)

   year                           energy_source twh_cons
1  1995                     biofuel_electricity     1.82
2  1995                        coal_electricity    24.18
3  1995                         gas_electricity     3.84
4  1995                       hydro_electricity    71.33
5  1995                     nuclear_electricity   377.23
6  1995                         oil_electricity    10.50
7  1995 other_renewable_exc_biofuel_electricity     0.51
8  1995                       solar_electricity     0.00
9  1995                        wind_electricity     0.00
10 2023                     biofuel_electricity     9.50
11 2023                        coal_electricity     2.16
12 2023                         gas_electricity    31.43
13 2023                       hydro_electricity    53.19
14 2023                     nuclear_electricity   335.65
15 2023                         oil_electricity     9.71
16 2023 other_renewable_exc_biofuel_electricity     0.60
17 2023                       solar_electricity    23.26
18 2023                        wind_electricity    48.61

print(edm_pivot_series(inpt_datf = cur_datf, time_col = 1, col_v = c(3)))

                            energy_source twh_cons_1995 twh_cons_2023
1                     biofuel_electricity          1.82          9.50
2                        coal_electricity         24.18          2.16
3                         gas_electricity          3.84         31.43
4                       hydro_electricity         71.33         53.19
5                     nuclear_electricity        377.23        335.65
6                         oil_electricity         10.50          9.71
7 other_renewable_exc_biofuel_electricity          0.51          0.60
8                       solar_electricity          0.00         23.26
9                        wind_electricity          0.00         48.61

print(datf)

    individual year                           energy_source twh_cons
1  France_1995 1995                     biofuel_electricity     1.82
2  France_1995 1995                        coal_electricity    24.18
3  France_1995 1995                         gas_electricity     3.84
4  France_1995 1995                       hydro_electricity    71.33
5  France_1995 1995                     nuclear_electricity   377.23
6  France_1995 1995                         oil_electricity    10.50
7  France_1995 1995 other_renewable_exc_biofuel_electricity     0.51
8  France_1995 1995                       solar_electricity     0.00
9  France_1995 1995                        wind_electricity     0.00
10 France_2023 2023                     biofuel_electricity     9.50
11 France_2023 2023                        coal_electricity     2.16
12 France_2023 2023                         gas_electricity    31.43
13 France_2023 2023                       hydro_electricity    53.19
14 France_2023 2023                     nuclear_electricity   335.65
15 France_2023 2023                         oil_electricity     9.71
16 France_2023 2023 other_renewable_exc_biofuel_electricity     0.60
17 France_2023 2023                       solar_electricity    23.26
18 France_2023 2023                        wind_electricity    48.61

print(edm_pivot_series(inpt_datf = cur_datf, time_col = 2, col_v = c(1, 4)))

  individual_1995                           energy_source twh_cons_1995
1     France_1995                     biofuel_electricity          1.82
2     France_1995                        coal_electricity         24.18
3     France_1995                         gas_electricity          3.84
4     France_1995                       hydro_electricity         71.33
5     France_1995                     nuclear_electricity        377.23
6     France_1995                         oil_electricity         10.50
7     France_1995 other_renewable_exc_biofuel_electricity          0.51
8     France_1995                       solar_electricity          0.00
9     France_1995                        wind_electricity          0.00
  individual_2023 twh_cons_2023
1     France_2023          9.50
2     France_2023          2.16
3     France_2023         31.43
4     France_2023         53.19
5     France_2023        335.65
6     France_2023          9.71
7     France_2023          0.60
8     France_2023         23.26
9     France_2023         48.61

}
