% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{elements_equalifier}
\alias{elements_equalifier}
\title{elements_equalifier}
\usage{
elements_equalifier(inpt_v, untl = 3)
}
\arguments{
\item{inpt_v}{is the input vector}

\item{untl}{is how many times each elements will be in the output vector}
}
\description{
Takes an input vector with elements that have different occurence, and output a vector with all these elements with the same number of occurence, see examples
}
\examples{

print(elements_equalifier(letters, untl = 2))

 [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
[20] "t" "u" "v" "w" "x" "y" "z" "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l"
[39] "m" "n" "o" "p" "q" "r" "s" "t" "u" "v" "w" "x" "y" "z"

print(elements_equalifier(c(letters, letters[-1]), untl = 2))

 [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p" "q" "r" "s"
[20] "t" "u" "v" "w" "x" "y" "z" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m"
[39] "n" "o" "p" "q" "r" "s" "t" "u" "v" "w" "x" "y" "z" "a"

}
