% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{elin_vec}
\alias{elin_vec}
\title{elin_vec}
\usage{
elin_vec(inpt_v, pattern_v, sep_ = "-")
}
\arguments{
\item{inpt_v}{is the input vector that may contains the sequence of elements contained in pattern_v}

\item{pattern_v}{is a vector containing the sequence of elements to search in inpt_v}

\item{sep_}{is a separator used in the background, make sure it does not correspond to any character in your input vector (inpt_v)}
}
\description{
Finds if a sequence of elements is present in a vector, if yes, the indexes of beginning and ending sequence in the vector is returned, see examples
}
\examples{

print(elin_vec(inpt_v = paste0(letters, 12), 
                  pattern_v = c("e12", "f12", "g12"), 
                  sep_ = "-"))

[1] 5 8

print(elin_vec(inpt_v = paste0(letters, 1), 
                  pattern_v = c("e1", "f1", "g1"), 
                  sep_ = "-"))

[1] 5 8

print(elin_vec(inpt_v = paste0(letters, 12), 
                  pattern_v = c("e1", "f1", "g1"), 
                  sep_ = "-"))

[1] NA 

}
