% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{elin_vec2}
\alias{elin_vec2}
\title{elin_vec2}
\usage{
elin_vec2(inpt_v, pattern_v, nvr_here = NA)
}
\arguments{
\item{inpt_v}{is the input vector that may contains the sequence of elements contained in pattern_v}

\item{pattern_v}{is a vector containing the sequence of elements to search in inpt_v}

\item{sep_}{is a separator used in the background, make sure it does not correspond to any character in your input vector (inpt_v)}
}
\description{
Finds if a sequence of elements is present in a vector, if yes, the indexes of beginning and ending sequence in the vector is returned, see examples. This uses an another algorithm than elin_vec1 function and allow for ReGeX expressions in pattern_v
}
\examples{

print(elin_vec2(inpt_v = paste0(letters, 1), 
                  pattern_v = c("e1", "f1", "g1"), 
                  nvr_here = NA))

[1] 5 8

print(elin_vec2(inpt_v = paste0(letters, 12), 
                  pattern_v = c("e1", "f1", "g1"), 
                  nvr_here = NA))

[1] 5 8

print(elin_vec2(inpt_v = paste0(letters, 12), 
                  pattern_v = c("^e1$", "^f1$", "^g1$"),
                  nvr_here = NA))

[1] NA

}
