% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{equalizer_v}
\alias{equalizer_v}
\title{equalizer_v}
\usage{
equalizer_v(inpt_v, depth = "max", default_val = "?")
}
\arguments{
\item{inpt_v}{is the input vector containing all the characters}

\item{depth}{is the depth parameter, defaults to "max" which means that it is equal to the character number of the element(s) in inpt_v that has the most}

\item{default_val}{is the default value that will be added to the output characters if those has an inferior length (characters) than the value of depth}
}
\description{
Takes a vector of character as an input and returns a vector with the elements at the same size. The size can be chosen via depth parameter.
}
\examples{

 print(equalizer_v(inpt_v=c("aa", "zzz", "q"), depth=2))

 #[1] "aa" "zz" "q?"

 print(equalizer_v(inpt_v=c("aa", "zzz", "q"), depth=12))

 #[1] "aa??????????" "zzz?????????" "q???????????"

}
