% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{extract_normal}
\alias{extract_normal}
\title{extract_normal}
\usage{
extract_normal(
  inpt_datf,
  mean,
  sd,
  accuracy,
  round_value = 1,
  normalised = FALSE,
  n = NA,
  tries = 3
)
}
\arguments{
\item{inpt_datf}{is the input dataset as a dataframe, values/modalities are in the first column and frequency (not normalised) is in the second column}

\item{mean}{is the mean of the target normal distribution}

\item{sd}{is the standard deviation of the target normal distribution}

\item{accuracy}{is how much of a difference beetween the points of the targeted normal distribution and the actual points is tolerated}

\item{round_value}{is the round value for the normal distribution used under the hood to compare the dataset and extract the best points, defaults to 1}

\item{normalised}{is if the input frequency is divided by n, if TRUE the parameter \code{n} must be filled}

\item{n}{is the number of points}

\item{tries}{is how many normal distributions are used under the hood to compare their points to the those in the input dataset, defaults to 3. The higher it is, the higher the number of different points from the input dataset will be in accordance for the normal distribution the function tries to build from the dataset. It does not increase by a lot but can be non-negligible and note that the higher the number of tries is, the higher the execution time of the function will be.}
}
\description{
Allow to extract values that fits a normal distribution from any kind of dataset, see examples and parameters
}
\examples{

sample_val <- round(rnorm(n = 72000, mean = 12, sd = 2), 1)
sample_freq <- unique_total(sample_val)
sample_qual <- infinite_char_seq(n = length(sample_freq))
datf_test <- data.frame(sample_qual, sample_freq)
n <- nrow(datf_test)
print(datf_test)
 
   sample_qual sample_freq
1             a          72
2             b        1155
3             c        1255
4             d         743
5             e         696
6             f        1028
7             g        1160
8             h        1219
9             i        1353
10            j        1336
11            k        1308
12            l         485
13            m        1306
14            n        1429
15            o         623
16            p        1172
17            q        1054
18            r         999
19            s         125
20            t        1461
21            u        1430
22            v         341
23            w        1453
24            x         427
25            y         869
26            z        1395
27           aa         841
28           ab         952
29           ac         246
30           ad         468
31           ae         237
32           af         555
33           ag        1297
34           ah         571
35           ai         349
36           aj         773
37           ak        1086
38           al        1281
39           am        1471
40           an        1236
41           ao         394
42           ap        1433
43           aq        1328
44           ar         976
45           as         640
46           at         308
47           au         698
48           av         864
49           aw        1346
50           ax        1349
51           ay           6
52           az        1071
53           ba         248
54           bb         929
55           bc         925
56           bd         452
57           be         207
58           bf         546
59           bg          62
60           bh         107
61           bi        1184
62           bj         739
63           bk         624
64           bl         850
65           bm        1408
66           bn         620
67           bo         202
68           bp          10
69           bq         700
70           br         397
71           bs        1291
72           bt         178
73           bu         397
74           bv        1089
75           bw        1301
76           bx         328
77           by        1348
78           bz          97
79           ca        1452
80           cb           4
81           cc         100
82           cd         593
83           ce         503
84           cf         164
85           cg          32
86           ch         259
87           ci        1089
88           cj         249
89           ck         165
90           cl          42
91           cm         143
92           cn         467
93           co         347
94           cp         143
95           cq          69
96           cr          18
97           cs         290
98           ct          55
99           cu         141
100          cv          86
101          cw         303
102          cx          88
103          cy          16
104          cz         213
105          da           3
106          db          75
107          dc          32
108          dd          66
109          de         105
110          df          34
111          dg          56
112          dh          17
113          di          22
114          dj         120
115          dk          54
116          dl           9
117          dm           8
118          dn          36
119          do          20
120          dp          26
121          dq          54
122          dr           8
123          ds          10
124          dt           4
125          du          53
126          dv          29
127          dw           1
128          dx           8
129          dy          10
130          dz           4
131          ea          22
132          eb           9
133          ec          17
134          ed          55
135          ee          21
136          ef           6
137          eg           4
138          eh           3
139          ei           7
140          ej           1
141          ek           4
142          el           2
143          em           5
144          en           4
145          eo           1
146          ep           2
147          eq           3
148          er           8
149          es           4
150          et           3
151          eu           3
152          ev           2
153          ew           2
154          ex           2
155          ey           1
156          ez           2
157          fa           2
158          fb           1

teste <- extract_normal(inpt_datf = datf_test,
                     mean = 10,
                     sd = 2,
                     accuracy = .1,
                     round_value = 1,
                     normalised = FALSE,
                     tries = 5)

print(length(unique(teste[, 1])) / n)

[1] 0.2848101 # so nearly 28.5 \% of the different points were in 
 #accordance with the construction of the target normal distribution

print(teste)

   values    frequency
1       dw 0.0001406866
2       dw 0.0001406866
3       dw 0.0001406866
4       el 0.0002813731
5       el 0.0002813731
6       el 0.0002813731
7       el 0.0002813731
8       da 0.0004220597
9       da 0.0004220597
10      cb 0.0005627462
11      cb 0.0005627462
12      em 0.0007034328
13      ay 0.0008441193
14      ay 0.0008441193
15      ei 0.0009848059
16      ei 0.0009848059
17      ei 0.0009848059
18      dm 0.0011254924
19      bp 0.0014068655
20      cy 0.0022509848
21      cy 0.0022509848
22      cy 0.0022509848
23      dh 0.0023916714
24      dh 0.0023916714
25      cr 0.0025323579
26      ee 0.0029544176
27      di 0.0030951041
28      dp 0.0036578503
29      dp 0.0036578503
30      cg 0.0045019696
31      cg 0.0045019696
32      df 0.0047833427
33      dn 0.0050647158
34      cl 0.0059088351
35      cl 0.0059088351
36      du 0.0074563872
37      du 0.0074563872
38      dg 0.0078784468
39      dg 0.0078784468
40      bg 0.0087225661
41      bg 0.0087225661
42      dd 0.0092853123
43      cq 0.0097073720
44      cq 0.0097073720
45       a 0.0101294316
46      cv 0.0120990433
47      cx 0.0123804164
48      cx 0.0123804164
49      bz 0.0136465954
50      cc 0.0140686550
51      bh 0.0150534609
52      bh 0.0150534609
53      dj 0.0168823860
54       s 0.0175858188
55       s 0.0175858188
56      cm 0.0201181767
57      cf 0.0230725943
58      ck 0.0232132808
59      bt 0.0250422060
60      bt 0.0250422060
61      be 0.0291221159
62      be 0.0291221159
63      cz 0.0299662352
64      cz 0.0299662352
65      be 0.0291221159
66      bo 0.0284186832
67      bt 0.0250422060
68      ck 0.0232132808
69      ck 0.0232132808
70      cm 0.0201181767
71      cu 0.0198368036
72       s 0.0175858188
73      dj 0.0168823860
74      bh 0.0150534609
75      bh 0.0150534609
76      de 0.0147720878
77      bz 0.0136465954
78      bz 0.0136465954
79      cx 0.0123804164
80      cv 0.0120990433
81      db 0.0105514913
82       a 0.0101294316
83      cq 0.0097073720
84      dd 0.0092853123
85      dd 0.0092853123
86      bg 0.0087225661
87      bg 0.0087225661
88      dg 0.0078784468
89      dk 0.0075970737
90      du 0.0074563872
91      cl 0.0059088351
92      cl 0.0059088351
93      dn 0.0050647158
94      df 0.0047833427
95      df 0.0047833427
96      cg 0.0045019696
97      dv 0.0040799100
98      dp 0.0036578503
99      di 0.0030951041
100     di 0.0030951041
101     ee 0.0029544176
102     cr 0.0025323579
103     dh 0.0023916714
104     cy 0.0022509848
105     cy 0.0022509848
106     cy 0.0022509848
107     cy 0.0022509848
108     dl 0.0012661790
109     dm 0.0011254924
110     ei 0.0009848059
111     ei 0.0009848059
112     ay 0.0008441193
113     ay 0.0008441193
114     em 0.0007034328
115     em 0.0007034328
116     cb 0.0005627462
117     cb 0.0005627462
118     da 0.0004220597
119     da 0.0004220597
120     el 0.0002813731
121     el 0.0002813731
122     el 0.0002813731
123     el 0.0002813731
124     dw 0.0001406866
125     dw 0.0001406866
126     dw 0.0001406866

}
