% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{extrm_dates}
\alias{extrm_dates}
\title{extrm_dates}
\usage{
extrm_dates(inpt_l, extrm = "min", sep = "-")
}
\arguments{
\item{inpt_l}{is the input vector}

\item{extrm}{is either "min" or "max", defaults to "min"}

\item{sep}{is the separator of the dates, defaults to "-"}
}
\description{
Allow to find the minimum or the maximum of a date in a vector. The format of dates is Year/Month/Day.
}
