% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{fillr}
\alias{fillr}
\title{fillr}
\usage{
fillr(inpt_v, ptrn_fill = "\\\\.\\\\.\\\\.\\\\d")
}
\arguments{
\item{inpt_v}{is the input vector}

\item{ptrn_fill}{is the pattern used to detect where the function has to fill the vector by the last element n times. It defaults to "...\\d" where "\\d" is the regex for an int value. So this paramater has to have "\\d" which designates n.}
}
\description{
Allow to fill a vector by the last element n times
}
\examples{

print(fillr(c("a", "b", "...3", "c")))

#[1] "a" "b" "b" "b" "b" "c"

}
