% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{fittr_v}
\alias{fittr_v}
\title{fittr_v}
\usage{
fittr_v(f_v, w_v, nvr_here = NA)
}
\arguments{
\item{f_v}{is the input vector}

\item{w_v}{is the vector containing the elements that can be in f_v}
}
\description{
Return the indexes of elements contained in "w_v" according to "f_v"
}
\examples{
print(fittr_v(f_v=c("non", "non", "non", "oui"), w_v=c("oui", "non", "non")))

[1] 4 1 2
}
