% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{fixer_nest_v}
\alias{fixer_nest_v}
\title{fixer_nest_v}
\usage{
fixer_nest_v(cur_v, pttrn_v, wrk_v)
}
\description{
Retur the elements of a vector "wrk_v" (1) that corresponds to the pattern of elements in another vector "cur_v" (2) according to another vector "pttrn_v" (3) that contains the patterof eleemnts.
}
\examples{
print(fixer_nest_v(cur_v=c("oui", "non", "peut-etre", "oui", "non", "peut-etre"), pttrn_v=c("oui", "non", "peut-etre"), 
                  wrk_v=c(1, 2, 3, 4, 5, 6)))

[1] 1 2 3 4 5 6

print(fixer_nest_v(cur_v=c("oui", "non", "peut-etre", "oui", "non", "peut-etre"), pttrn_v=c("oui", "non"), 
                  wrk_v=c(1, 2, 3, 4, 5, 6)))

[1]  1  2 NA  4  5 NA
}
