% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{fold_rec}
\alias{fold_rec}
\title{fold_rec}
\usage{
fold_rec(xmax, xmin = 1, pathc = ".")
}
\arguments{
\item{xmax}{is the end depth value}

\item{xmin}{is the start depth value}

\item{pathc}{is the reference path}
}
\description{
Allow to get all the files recursively from a path according to an end and start depth value. If you want to have an other version of this function that uses a more sophisticated algorythm (which can be faster), check file_rec2. Depth example: if i have dir/dir2/dir3, dir/dir2b/dir3b, i have a depth equal to 3
}
