% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{format_date}
\alias{format_date}
\title{format_date}
\usage{
format_date(f_dialect, sentc, sep_in = "-", sep_out = "-")
}
\arguments{
\item{f_dialect}{are the months from the language of which the month come}

\item{sentc}{is the date to convert}

\item{sep_in}{is the separator of the dat input (default is "-")}

\item{sep_out}{is the separator of the converted date (default is "-")}
}
\description{
Allow to convert xx-month-xxxx date type to xx-xx-xxxx
}
\examples{

print(format_date(f_dialect=c("janvier", "février", "mars", "avril", "mai", "juin",
"juillet", "aout", "septembre", "octobre", "novembre", "décembre"), sentc="11-septembre-2023"))

#[1] "11-09-2023"

}
