% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{get_rec}
\alias{get_rec}
\title{get_rec}
\usage{
get_rec(pathc = ".")
}
\arguments{
\item{pathc}{is the reference path
example: if i have dir/dir2/dir3, dir/dir2b/dir3b, i have a depth equal to 3}
}
\description{
Allow to get the value of directorie depth from a path.
}
