% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{glue_groupr_v}
\alias{glue_groupr_v}
\title{glue_groupr_v}
\usage{
glue_groupr_v(inpt_v, group_v = c(), untl)
}
\arguments{
\item{inpt_v}{is the input vector}

\item{is}{a vector containing all the elements that will be glued in the output vector}
}
\description{
Takes an input vector and returns the same vector unlike that certain elements will be glued as an unique element according to thoses designated in a special vector, see examples.
}
\examples{

print(glue_groupr_v(inpt_v = c("o", "-", "-", "u", "i", "-", "n", 
 "o", "-", "-", "-", "zz", "/", "/"), group_v = c("-", "/")))

[1] "o"   "--"  "u"   "i"   "-"   "n"   "o"   "---" "zz"  "//" 

print(glue_groupr_v(inpt_v = c("o", "-", "-", "u", "i", "-", "n", 
 "o", "-", "-", "-", "-", "zz", "/", "/"), group_v = c("-", "/"), untl = 3))

[1] "o"   "--"  "u"   "i"   "-"   "n"   "o"   "---" "-"   "zz"  "//"  

print(glue_groupr_v(inpt_v = c("o", "-", "-", "u", "i", "-", "n", 
"o", "-", "-", "-", "-", "zz", "/", "/"), group_v = c("-", "/"), untl = 2))

[1] "o"  "--" "u"  "i"  "-"  "n"  "o"  "--" "--" "zz" "//" 

}
