% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{groupr_datf}
\alias{groupr_datf}
\title{groupr_datf}
\usage{
groupr_datf(
  inpt_datf,
  condition_lst,
  val_lst,
  conjunction_lst,
  rtn_val_pos = c()
)
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{condition_lst}{is a list containing all the condition as a vector for each group}

\item{val_lst}{is a list containing all the values associated with condition_lst as a vector for each group}

\item{conjunction_lst}{is a list containing all the conjunctions associated with condition_lst and val_lst as a vector for each group}

\item{rtn_val_pos}{is a vector containing all the group flag value like this ex: c("flag1", "flag2", "flag3")}
}
\description{
Allow to create groups from a dataframe. Indeed, you can create conditions that lead to a flag value for each cell of the input dataframeaccording to the cell value. This function is based on see_datf and nestr_datf2 functions.
}
\examples{
interactive()

datf1 <- data.frame(c(1, 2, 1), c(45, 22, 88), c(44, 88, 33))
                                                                      
val_lst <- list(list(c(1), c(1)), list(c(2)), list(c(44, 88)))

condition_lst <- list(c(">", "<"), c("\%\%"), c("==", "=="))

conjunction_lst <- list(c("|"), c(), c("|"))

rtn_val_pos <- c("+", "++", "+++")

print(groupr_datf(inpt_datf=datf1, val_lst=val_lst, condition_lst=condition_lst, 
conjunction_lst=conjunction_lst, rtn_val_pos=rtn_val_pos))

#    X1  X2  X3
#1 <NA>   + +++
#2   ++  ++ +++
#3 <NA> +++   +

}
