% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{gsub_mult}
\alias{gsub_mult}
\title{gsub_mult}
\usage{
gsub_mult(inpt_v, pattern_v = c(), replacement_v = c())
}
\arguments{
\item{inpt_v}{is a vector containing all the elements that contains expressions to be substituted}

\item{pattern_v}{is a vector containing all the patterns to be substituted in any elements of inpt_v}

\item{replacement_v}{is a vector containing the expression that are going to substituate those provided by pattern_v}
}
\description{
Performs a gsub operation with n patterns and replacements.
}
\examples{

print(gsub_mult(inpt_v = c("X and Y programming languages are great", "More X, more X!"), 
               pattern_v = c("X", "Y", "Z"), 
               replacement_v = c("C", "R", "GO")))
[1] "C and R programming languages are great"
[2] "More C, more C!"                        

}
