% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{id_keepr}
\alias{id_keepr}
\title{id_keepr_datf}
\usage{
id_keepr(inpt_datf, col_v = c(), el_v = c(), rstr_l = NA)
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{col_v}{is the vector containing the column numbers or names to be compared to their respective elements in "el_v"}

\item{el_v}{is a vector containing the elements that may be contained in their respective column described in "col_v"}

\item{rstr_l}{is a list containing the vector composed of the indexes of the elements chosen for each comparison. If the length of the list is inferior to the lenght of comparisons, so the last vector of rstr_l will be the same as the last one to fill make rstr_l equal in term of length to col_v and el_v}
}
\description{
Allow to get the original indexes after multiple equality comparaison according to the original number of row
}
\examples{

datf1 <- data.frame(c("oui", "oui", "oui", "non", "oui"), 
     c("opui", "op", "op", "zez", "zez"), c(5:1), c(1:5))

print(id_keepr(inpt_datf=datf1, col_v=c(1, 2), el_v=c("oui", "op")))

#[1] 2 3

print(id_keepr(inpt_datf=datf1, col_v=c(1, 2), el_v=c("oui", "op"), 
     rstr_l=list(c(1:5), c(3, 2, 2, 2, 3))))

#[1] 2 3

print(id_keepr(inpt_datf=datf1, col_v=c(1, 2), el_v=c("oui", "op"), 
     rstr_l=list(c(1:5), c(3))))

#[1] 3

print(id_keepr(inpt_datf=datf1, col_v=c(1, 2), el_v=c("oui", "op"), rstr_l=list(c(1:5))))

#[1] 2 3

}
