% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{incr_fillr}
\alias{incr_fillr}
\title{incr_fillr}
\usage{
incr_fillr(inpt_v, wrk_v = NA, default_val = NA, step = 1)
}
\arguments{
\item{inpt_v}{is the asending double only composed vector}

\item{wrk_v}{is the other vector (size equal to inpt_v), defaults to NA}

\item{default_val}{is the default value put when the difference between two following elements of inpt_v is greater than step, defaults to NA}

\item{step}{is the allowed difference between two elements of inpt_v}
}
\description{
Take a vector uniquely composed by double and sorted ascendingly, a step, another vector of elements whose length is equal to the length of the first vector, and a default value. If an element of the vector is not equal to its predecessor minus a user defined step, so these can be the output according to the parameters (see example):
}
\examples{

print(incr_fillr(inpt_v=c(1, 2, 4, 5, 9, 10), 
                wrk_v=NA, 
                default_val="increasing"))

#[1]  1  2  3  4  5  6  7  8  9 10

print(incr_fillr(inpt_v=c(1, 1, 2, 4, 5, 9), 
                wrk_v=c("ok", "ok", "ok", "ok", "ok"), 
                default_val=NA))

#[1] "ok" "ok" "ok" NA   "ok" "ok" NA   NA   NA  

print(incr_fillr(inpt_v=c(1, 2, 4, 5, 9, 10), 
                wrk_v=NA, 
                default_val="NAN"))

#[1] "1"   "2"   "NAN" "4"   "5"   "NAN" "NAN" "NAN" "9"   "10" 

}
