% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{infinite_char_seq}
\alias{infinite_char_seq}
\title{infinite_char_seq}
\usage{
infinite_char_seq(n, base_char = letters)
}
\arguments{
\item{n}{is how many sequence of numbers will be generated}

\item{base_char}{is the vector containing the elements from which the sequence is generated}
}
\description{
Allow to generate an infinite sequence of unique letters
}
\examples{

print(infinite_char_seq(28))

 [1] "a"  "b"  "c"  "d"  "e"  "f"  "g"  "h"  "i"  "j"  "k"  "l"  "m"  "n"  "o" 
[16] "p"  "q"  "r"  "s"  "t"  "u"  "v"  "w"  "x"  "y"  "a"  "aa" "ab"

}
