% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{inner_all}
\alias{inner_all}
\title{inner_all}
\usage{
inner_all(..., keep_val = FALSE, id_v)
}
\arguments{
\item{...}{are all the dataframes etc}

\item{keep_val}{is if you want to keep the id column}

\item{id_v}{is the common id of all the dataframes etc}
}
\description{
Allow to apply inner join on n dataframes, datatables, tibble
}
\examples{

datf1 <- data.frame(
        "id1"=c(1:5),
        "var1"=c("oui", "oui", "oui", "non", "non")
)

datf2 <- data.frame(
        "id1"=c(1, 2, 3, 7, 9),
        "var1"=c("oui2", "oui2", "oui2", "non2", "non2")
)

print(inner_all(datf1, datf2, keep_val=FALSE, id_v="id1"))

id1 var1.x var1.y
1   1    oui   oui2
2   2    oui   oui2
3   3    oui   oui2

}
