% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{intersect_all}
\alias{intersect_all}
\title{intersect_all}
\usage{
intersect_all(...)
}
\arguments{
\item{...}{is all the vector you want to calculate the intersection from}
}
\description{
Allows to calculate the intersection between n vectors
}
\examples{

print(intersect_all(c(1:5), c(1, 2, 3, 6), c(1:4)))

[1] 1 2 3

cur_lst <- list()
cur_lst <- append(x = cur_lst, values = list(c(1:10)))
cur_lst <- append(x = cur_lst, values = list(c(5:17)))
cur_lst <- append(x = cur_lst, values = list(c(-5:7)))

print(intersect_all(cur_lst))

[1] 5 6 7

}
