% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{intersect_mod2}
\alias{intersect_mod2}
\title{intersect_mod2}
\usage{
intersect_mod2(inpt_datf, inter_col, mod_col, n_min = 1)
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{inter_col}{is the column number or name of the elements that may be in common to each mod}

\item{mod_col}{is the column number or name of the mods}

\item{n_min}{is the minimum number of elements in common each mod have to have to be outputed in the returned dataframe}
}
\description{
Returns the mods that have elemnts in common, uses a different algorith than intersect_mod1, see examples
}
\examples{

datf <- data.frame("country" = c("Germany", "France", "Italy", "France", "Germany", "Italy"),
                   "year" = c(2000, 2000, 2001, 2002, 2003, 2004),
                   "random_value" = c(1, 3, 2, 1, 2, 2))
print(datf)

  country year random_value
1 Germany 2000            1
2  France 2000            3
3   Italy 2001            2
4  France 2002            1
5 Germany 2003            2
6   Italy 2004            2

print(intersect_mod2(inpt_datf = datf, inter_col = "year", mod_col = "country"))

  country year random_value
1 Germany 2000            1
2  France 2000            3

}
