% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{is_divisible}
\alias{is_divisible}
\title{is_divisible}
\usage{
is_divisible(inpt_v = c(), divisible_v = c())
}
\arguments{
\item{inpt_v}{is the input vector}

\item{divisible_v}{is the vector containing all the numbers that will try to divide those contained in inpt_v}
}
\description{
Takes a vector as an input and returns all the elements that are divisible by all choosen numbers from another vector.
}
\examples{

 print(is_divisible(inpt_v=c(1:111), divisible_v=c(2, 4, 5)))

 #[1]  20  40  60  80 100

}
