% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{join_n_lvl}
\alias{join_n_lvl}
\title{join_n_lvl}
\usage{
join_n_lvl(frst_datf, scd_datf, join_type = c(), lst_pair = list())
}
\arguments{
\item{frst_datf}{is the first data.frame (table)}

\item{scd_datf}{is the second data.frame (table)}

\item{join_type}{is a vector containing all the join type ("left", "inner", "right") for each variable}

\item{lst_pair}{is a lis of vectors. The vectors refers to a multi-level join. Each vector should have a length of 1. Each vector should have a name. Its name refers to the column name of multi-level variable and its value refers to the column name of the join variable.}
}
\description{
Allow to see the progress of the multi-level joins of the different variables modalities. Here, multi-level joins is a type of join that usually needs a concatenation of two or more variables to make a key. But here, there is no need to proceed to a concatenation. See examples.
}
\examples{

datf3 <- data.frame("vil"=c("one", "one", "one", "two", "two", "two"),
                     "charac"=c(1, 2, 2, 1, 2, 2),
                     "rev"=c(1250, 1430, 970, 1630, 2231, 1875),
                     "vil2" = c("one", "one", "one", "two", "two", "two"),
                     "idl2" = c(1:6))
datf4 <- data.frame("vil"=c("one", "one", "one", "two", "two", "three"),
                    "charac"=c(1, 2, 2, 1, 1, 2),
                     "rev"=c(1.250, 1430, 970, 1630, 593, 456),
                     "vil2" = c("one", "one", "one", "two", "two", "two"),
                     "idl2" = c(2, 3, 1, 5, 5, 5))

print(join_n_lvl(frst_datf=datf3, scd_datf=datf4, lst_pair=list(c("charac" = "vil"), c("vil2" = "idl2")), 
                 join_type=c("inner", "left")))

[1] "pair: charac vil"
|  |   0\%
1 
|= |  50\%
2 
|==| 100\%
[1] "pair: vil2 idl2"
|  |   0\%
one 
|= |  50\%
two 
|==| 100\%

  main_id.x vil.x charac.x rev.x vil2.x idl2.x main_id.y vil.y charac.y rev.y
1  1oneone1   one        1  1250    one      1      <NA>  <NA>       NA    NA
2  2oneone2   one        2  1430    one      2      <NA>  <NA>       NA    NA
3  2oneone3   one        2   970    one      3  2oneone3   one        2  1430
4  1twotwo4   two        1  1630    two      4      <NA>  <NA>       NA    NA
  vil2.y idl2.y
1   <NA>     NA
2   <NA>     NA
3    one      3
4   <NA>     NA

}
