% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{just_anything}
\alias{just_anything}
\title{just_anything}
\usage{
just_anything(inpt_v, symbol_ = "-", anything_v = c())
}
\arguments{
\item{inpt_v}{is the input vector}

\item{symbol_}{is the chosen symbol to replace numbers}
}
\description{
Extract only the letters from all elements of a vector, see examples
}
\examples{

print(just_anything(inpt_v = c("oui222jj644", "oui122jj"), 
symbol_ = "-", anything_v = letters))

[1] "oui-jj-" "oui-jj" 

}
