% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{just_anything3}
\alias{just_anything3}
\title{just_anything3}
\usage{
just_anything3(inpt_v, anything_v = c())
}
\arguments{
\item{inpt_v}{is the input vector}
}
\description{
Extract only the letters from all elements of a vector, see examples
}
\examples{

print(just_anything3(inpt_v = c("oui222jj644", "oui122jj"), 
 anything_v = letters))

[1] "ouijj" "ouijj"

}
