% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{just_chr3}
\alias{just_chr3}
\title{just_chr3}
\usage{
just_chr3(inpt_v)
}
\arguments{
\item{inpt_v}{is the input vector}
}
\description{
Extract only the letters from all elements of a vector, see examples
}
\examples{

print(just_chr3(inpt_v = c("oui222jj644", "oui122jj")))

[1] "ouijj" "ouijj"

}
