% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{just_nb}
\alias{just_nb}
\title{just_nb}
\usage{
just_nb(inpt_v, symbol_ = "-")
}
\arguments{
\item{inpt_v}{is the input vector}

\item{symbol_}{is the chosen symbol to replace numbers}
}
\description{
Extract only the letters from all elements of a vector, see examples
}
\examples{

print(just_nb(inpt_v = c("oui222jj644", "oui122jj"), 
    symbol_ = "-"))

[1] "-222-44" "-122-"  

}
