% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{just_nb2}
\alias{just_nb2}
\title{just_nb2}
\usage{
just_nb2(inpt_v, track_ = TRUE, symbol_ = "-")
}
\arguments{
\item{inpt_v}{is the input vector}

\item{symbol_}{is the chosen symbol to replace numbers}

\item{track_len}{is a boolean, if set to TRUE, the chosen symbol will equal the len of the missing numbers, see examples}
}
\description{
Extract only the letters from all elements of a vector, see examples
}
\examples{

print(just_nb2(inpt_v = c("oui222jj44", "oui122jj"), 
   track_ = TRUE, symbol_ = "-"))

[1] "---222--44" "---122--"  

}
