% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{just_nb3}
\alias{just_nb3}
\title{just_nb3}
\usage{
just_nb3(inpt_v)
}
\arguments{
\item{inpt_v}{is the input vector}
}
\description{
Extract only the letters from all elements of a vector, see examples
}
\examples{

print(just_nb3(inpt_v = c("oui222jj644", "oui122jj")))

[1] 222644 122

}
