% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{list_files}
\alias{list_files}
\title{list_files}
\usage{
list_files(patternc, pathc = ".")
}
\arguments{
\item{patternc}{is a vector containing all the exensions you want}

\item{pathc}{is the path, can be a vector of multiple path because list.files() supports it.}
}
\description{
A list.files() based function addressing the need of listing the files with extension a or or extension b ...
}
