% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{lst_flatnr}
\alias{lst_flatnr}
\title{lst_flatnr}
\usage{
lst_flatnr(inpt_l)
}
\arguments{
\item{inpt_l}{is the input list}
}
\description{
Flatten a list to a vector
}
\examples{

print(lst_flatnr(inpt_l=list(c(1, 2), c(5, 3), c(7, 2, 7))))

#[1] 1 2 5 3 7 2 7

}
