% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{match_by}
\alias{match_by}
\title{match_by}
\usage{
match_by(to_match_v = c(), inpt_v = c(), inpt_ids = c())
}
\arguments{
\item{to_match_v}{is the vector containing all the elements to match}

\item{inpt_v}{is the input vector containong all the elements that could contains the elements to match. Each elements is linked to an element from inpt_ids at any given index, see examples. So inpt_v and inpt_ids must be the same size}

\item{inpt_ids}{is the vector containing all the ids for the elements in inpt_v. An element is linked to the id x is both are at the same index. So inpt_v and inpt_ids must be the same size}
}
\description{
Allow to match elements by ids, see examples.
}
\examples{

print(match_by(to_match_v = c("a"), inpt_v = c("a", "z", "a", "p", "p", "e", "e", "a"), 
               inpt_ids = c(1, 1, 1, 2, 2, 3, 3, 3)))

[1] 1 8

print(match_by(to_match_v = c("a"), inpt_v = c("a", "z", "a", "a", "p", "e", "e", "a"), 
               inpt_ids = c(1, 1, 1, 2, 2, 3, 3, 3)))

[1] 1 4 8

print(match_by(to_match_v = c("a", "e"), inpt_v = c("a", "z", "a", "a", "p", "e", "e", "a"), 
               inpt_ids = c(1, 1, 1, 2, 2, 3, 3, 3)))

[1] 1 4 8 6

}
