% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{match_n}
\alias{match_n}
\title{match_n}
\usage{
match_n(vec, mc, n = 1, wnb = "#####")
}
\arguments{
\item{vec}{is th einput vector
mc is a vector containing the values you want to get the index for the nth occurence in vec}

\item{wnb}{is a string you are sure is not in mc}

\item{is}{the value of the occurence}
}
\description{
Allow to get the indexes for the nth occurence of a value in a vector. Example: c(1, 2, 3, 1, 2), the first occurence of 1 and 2 is at index 1 and 2 respectively, but the second occurence is respectively at the 4th and 5th index.
}
