% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{match_na_omit}
\alias{match_na_omit}
\title{match_na_omit}
\usage{
match_na_omit(x, table)
}
\arguments{
\item{x}{is the vector of the patterns to be matched}

\item{table}{is the vector that may contain the patterns to be matched}
}
\description{
Performs a match, but remove the NA values in the output if there is one or many, see examples.
}
\examples{

match_na_omit(x = c("oui", "non", "2"), table = c("1", "oui", "oui", "ee", "non"))

[1] 2 5

}
