% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{multitud}
\alias{multitud}
\title{multitud}
\usage{
multitud(l, sep_ = "")

multitud(l, sep_ = "")
}
\arguments{
\item{l}{is the list}

\item{sep_}{is the separator between elements (default is set to "" as you see in the example)}
}
\description{
From a list containing vectors allow to generate a vector following this rule:

From a list containing vectors allow to generate a vector following this rule:
}
\details{
list(c("a", "b"), c("1", "2"), c("A", "Z", "E")) --> c("a1A", "a2A", "b1A", "b2A", "a1Z", ...)

list(c("a", "b"), c("1", "2"), c("A", "Z", "E")) --> c("a1A", "a2A", "b1A", "b2A", "a1Z", ...)
}
