% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{mutate_vector}
\alias{mutate_vector}
\title{mutate_vector}
\usage{
mutate_vector(
  inpt_v,
  n_inpt,
  nvr_here = "NULL",
  base_seed = "random_data2.csv"
)
}
\arguments{
\item{inpt_v}{is the input vector containing the elements to evenly select}

\item{n_inpt}{is the number of elements of the output vector}

\item{nvr_here}{is a known value that is never in 'inpt_v'}

\item{base_seed}{is a csv filename containing the values for the reimplentation of the uniform function to work with}
}
\description{
Allow to select elements from a vector according to a uniform distribution (better than runif, especially for small vectors). You can choose the seed that the reimplementation of the uniform distribution will work with, defaults to 'random_data2.csv', see examples
}
\examples{

library("edm1")
library("ggplot2")

test_v <- letters

x <- mutate_vector(inpt_v = test_v, 
                   n_inpt = 50000, 
                   nvr_here = "NULL", 
                   base_seed = "random_data2.csv")

length(x)

[1] 5000

datf <- occu(x)

datf \%>\%
  ggplot(mapping = aes(x = var, y = occurence)) +
  geom_col() +
  theme_minimal()

}
