% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{nb2_follow}
\alias{nb2_follow}
\title{nb2_follow}
\usage{
nb2_follow(inpt_v, inpt_idx, inpt_follow_v = c())
}
\arguments{
\item{inpt_v}{is the input vector}

\item{inpt_idx}{is the index}

\item{inpt_follow_v}{is a vector containing the patterns that are potentially just after inpt_nb}
}
\description{
Allows to get the number and pattern of potential continuous pattern after an index of a vector, see examples
}
\examples{

print(nb2_follow(inpt_v = c(1:12), inpt_idx = 4, inpt_follow_v = c(5)))

[1] 1 5

print(nb2_follow(inpt_v = c(1, "non", "oui", "oui", "oui", "nop", 5), inpt_idx = 2, inpt_follow_v = c("5", "oui")))

[1] "3"   "oui"

}
