% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{nb_decimal}
\alias{nb_decimal}
\title{nb_decimal}
\usage{
nb_decimal(inpt_x)
}
\arguments{
\item{inpt_x}{is the input number}
}
\description{
Returns the number of deciaml of the input number.
}
\examples{

print(nb_decimal(inpt_x = "22.4356"))

[1] 4

print(nb_decimal(inpt_x = "2.456"))

[1] 3

}
