% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{nb_follow}
\alias{nb_follow}
\title{nb_follow}
\usage{
nb_follow(inpt_v, inpt_idx, inpt_follow_v = c())
}
\arguments{
\item{inpt_v}{is the input vector}

\item{inpt_idx}{is the index}

\item{inpt_follow_v}{is a vector containing all the potential patterns that may follow the element in the vector at the index inpt_idx}
}
\description{
Allow to get the number of certains patterns that may be after an index of a vector, see examples
}
\examples{

print(nb_follow(inpt_v = c(1:13), inpt_idx = 6, inpt_follow_v = c(5:9)))

[1] 3

print(nb_follow(inpt_v = c("ou", "nn", "pp", "zz", "zz", "ee", "pp"), inpt_idx = 2, 
                inpt_follow_v = c("pp", "zz")))

[1] 3

}
