% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{nest_v}
\alias{nest_v}
\title{nest_v}
\usage{
nest_v(f_v, t_v, step = 1, after = 1)
}
\arguments{
\item{f_v}{is the vector that will welcome the nested vector t_v}

\item{t_v}{is the imbriquator vector}

\item{step}{defines after how many elements of f_v the next element of t_v can be put in the output}

\item{after}{defines after how many elements of f_v, the begining of t_v can be put}
}
\description{
Nest two vectors according to the following parameters.
}
\examples{

print(nest_v(f_v=c(1, 2, 3, 4, 5, 6), t_v=c("oui", "oui2", "oui3", "oui4", "oui5", "oui6"), 
     step=2, after=2))

#[1] "1"    "2"    "oui"  "3"    "4"    "oui2" "5"    "6"    "oui3" "oui4"

}
