% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{nestr_datf1}
\alias{nestr_datf1}
\title{nestr_datf1}
\usage{
nestr_datf1(
  inptf_datf,
  inptt_pos_datf,
  nestr_datf,
  yes_val = TRUE,
  inptt_neg_datf = NA
)
}
\arguments{
\item{inptf_datf}{is the input dataframe (1b)}

\item{inptt_pos_datf}{is the dataframe (2b) that corresponds to the (1a) values}

\item{nestr_datf}{is the dataframe (2b) that has the special value (2a)}

\item{yes_val}{is the special value (2a)}

\item{inptt_neg_datf}{is the dataframe (4b) that has the (3a) values, defaults to NA}
}
\description{
Allow to write a value (1a) to a dataframe (1b) to its cells that have the same coordinates (row and column) than the cells whose value is equal to a another special value (2a), from another another dataframe (2b). The value (1a) depends of the cell  value coordinates of the third dataframe (3b). If a cell coordinates (1c) of the first dataframe (1b) does not correspond to the coordinates of a good returning cell value (2a) from the dataframe (2b), so this cell (1c) can have its value changed to the same cell coordinates value (3a) of a third dataframe (4b), if (4b) is not set to NA.
}
\examples{

print(nestr_datf1(inptf_datf=data.frame(c(1, 2, 1), c(1, 5, 7)), 
inptt_pos_datf=data.frame(c(4, 4, 3), c(2, 1, 2)), 
inptt_neg_datf=data.frame(c(44, 44, 33), c(12, 12, 12)), 
nestr_datf=data.frame(c(TRUE, FALSE, TRUE), c(FALSE, FALSE, TRUE)), yes_val=TRUE)) 

#  c.1..2..1. c.1..5..7.
#1          4         12
#2         44         12
#3          3          2

print(nestr_datf1(inptf_datf=data.frame(c(1, 2, 1), c(1, 5, 7)), 
inptt_pos_datf=data.frame(c(4, 4, 3), c(2, 1, 2)), 
inptt_neg_datf=NA, 
nestr_datf=data.frame(c(TRUE, FALSE, TRUE), c(FALSE, FALSE, TRUE)), yes_val=TRUE))

#   c.1..2..1. c.1..5..7.
#1          4          1
#2          2          5
#3          3          2

}
