% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{nestr_datf2}
\alias{nestr_datf2}
\title{nestr_datf2}
\usage{
nestr_datf2(inptf_datf, rtn_pos, rtn_neg = NA, nestr_datf, yes_val = T)
}
\arguments{
\item{inptf_datf}{is the input dataframe (1b)}

\item{rtn_pos}{is the special value (1a)}

\item{rtn_neg}{is the special value (3a)}

\item{nestr_datf}{is the dataframe (2b)}

\item{yes_val}{is the special value (2a)}
}
\description{
Allow to write a special value (1a) in the cells of a dataframe (1b) that correspond (row and column) to whose of another dataframe (2b) that return another special value (2a). The cells whose coordinates do not match the coordinates of the dataframe (2b), another special value can be written (3a) if not set to NA.
}
\examples{

print(nestr_datf2(inptf_datf=data.frame(c(1, 2, 1), c(1, 5, 7)), rtn_pos="yes", 
rtn_neg="no", nestr_datf=data.frame(c(TRUE, FALSE, TRUE), c(FALSE, FALSE, TRUE)), yes_val=TRUE)) 

#  c.1..2..1. c.1..5..7.
#1        yes         no
#2         no         no
#3        yes        yes

}
