% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{nestr_df2}
\alias{nestr_df2}
\title{nestr_df2}
\usage{
nestr_df2(inptf_df, rtn_pos, rtn_neg = NA, nestr_df, yes_val = T)
}
\arguments{
\item{rtn_pos}{is the special value (1a)}

\item{rtn_neg}{is the special value (3a)}

\item{nestr_df}{is the dataframe (2b)}

\item{yes_val}{is the special value (2a)}

\item{inpt_df}{is the input dataframe (1b)}
}
\description{
Allow to write a special value (1a) in the cells of a dataframe (1b) that correspond (row and column) to whose of another dataframe (2b) that return another special value (2a). The cells whose coordinates do not match the coordinates of the dataframe (2b), another special value can be written (3a) if not set to NA.
}
\examples{
nestr_df2(inpt_df=data.fame(c(1, 2, 1), c(1, 5, 7)), rtn_pos="yes", 
rtn_neg="no", nestr_df=data.frame(c(T, F, T), c(F, F, T), yes_val=T)) 
}
