% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{new_group2}
\alias{new_group2}
\title{new_group2}
\usage{
new_group2(inpt_datf, lst_grp = list(), grp_col)
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{lst_grp}{is a list containing vectors containing some of the existing groups to create the new groups, see example}

\item{grp_col}{is the column name or number of the existing groups}
}
\description{
Allow to create new dataframe groups based on existing groups, see example.
}
\examples{

datf <- data.frame("grp" = c(rep(x = "a", times = 12),
                             rep(x = "b", times = 4),
                             rep(x = "c", times = 10)))
print(new_group1(inpt_datf = datf,
        grp_col = "grp",
        lst_grp = list(c("a", "b"), "c")
         )
       )

   grp new_group
1    a         1
2    a         1
3    a         1
4    a         1
5    a         1
6    a         1
7    a         1
8    a         1
9    a         1
10   a         1
11   a         1
12   a         1
13   b         1
14   b         1
15   b         1
16   b         1
17   c         2
18   c         2
19   c         2
20   c         2
21   c         2
22   c         2
23   c         2
24   c         2
25   c         2
26   c         2

}
