% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{non_unique}
\alias{non_unique}
\title{non_unique}
\usage{
non_unique(inpt_v, occu = ">-1-")
}
\arguments{
\item{inpt_v}{is the input vector containing the elements}

\item{occu}{is a parameter that specifies the occurence of the elements that must be returned, defaults to ">-1-" it means that the function will return all the elements that are present more than one time in inpt_v. The synthax is the following "comparaison_type-actual_value-". The comparaison type may be "==" or ">". Occu can also be a vector containing all the occurence that must have the elements to be returned.}
}
\description{
Returns the element that are not unique from the input vector
}
\examples{

print(non_unique(inpt_v=c("oui", "oui", "non", "non", "peut", "peut1", "non")))

#[1] "oui" "non"

print(non_unique(inpt_v=c("oui", "oui", "non", "non", "peut", "peut1", "non"), occu="==-2-"))

#[1] "oui"

print(non_unique(inpt_v=c("oui", "oui", "non", "non", "peut", "peut1", "non"), occu=">-2-"))

#[1] "non"

print(non_unique(inpt_v=c("oui", "oui", "non", "non", "peut", "peut1", "non"), occu=c(1, 3)))

#[1] "non"   "peut"  "peut1"

}
