% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{normal_offset_prob}
\alias{normal_offset_prob}
\title{normal_offset_prob}
\usage{
normal_offset_prob(inpt_v = c(), mean_inpt, sd_inpt)
}
\arguments{
\item{mean_inpt}{is the mean of the normal distribution}

\item{sd_inpt}{is the standard deviations of the normal distribution}
}
\description{
Returns the probability of the most offset value in the input vector accrding to the normal distribution, see examples.
}
\examples{

print(normal_offset_prob(inpt_v = x,
                  mean_inpt = 15,
                  sd_inpt = 1))

[1] 0.01146424

print(normal_offset_prob(inpt_v = x,
                  mean_inpt = 15,
                  sd_inpt = 2))

[1] 0.08212756

}
