% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{old_to_new_idx}
\alias{old_to_new_idx}
\title{old_to_new_idx}
\usage{
old_to_new_idx(inpt_v = c())
}
\arguments{
\item{inpt_v}{is the input vector}
}
\description{
Allow to convert index of elements in a vector \code{inpt_v} to index of an vector type 1:sum(nchar(inpt_v)), see examples
}
\examples{

print(old_to_new_idx(inpt_v = c("oui", "no", "eeee")))

[1] 1 1 1 2 2 3 3 3 3

}
