% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{or_bool2}
\alias{or_bool2}
\title{or_bool2}
\usage{
or_bool2(inpt_datf)
}
\description{
Returns a boolean vector according to 'or' condition on boolean vector in a dataframe, each column represents a boolean vector, see examples, uses a different algorithm than or_bool1
}
\examples{

print(or_bool2(inpt_datf = data.frame("bool1" = c(TRUE, FALSE, FALSE, TRUE),
                                       "bool2" = c(TRUE, TRUE, FALSE, TRUE))))

[1]  TRUE  TRUE FALSE  TRUE

}
