% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{pairs_findr}
\alias{pairs_findr}
\title{pairs_findr}
\usage{
pairs_findr(inpt, ptrn1 = "(", ptrn2 = ")")
}
\arguments{
\item{inpt}{is the input character}

\item{ptrn1}{is the first pattern ecountered in the pair}

\item{ptrn2}{is the second pattern in the pair}
}
\description{
Takes a character as input and detect the pairs of pattern, like the parenthesis pais if the pattern is "(" and then ")"
}
\examples{

print(pairs_findr(inpt="ze+(yu*45/(jk+zz)*(o()p))-(re*(rt+qs)-fg)"))

[[1]]
 [1] 4 1 1 3 2 2 3 4 6 5 5 6

[[2]]
 [1]  4 11 17 19 21 22 24 25 27 31 37 41

}
