% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{paste_datf2}
\alias{paste_datf2}
\title{paste_datf2}
\usage{
paste_datf2(inpt_datf, sep = "")
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{sep}{is the separator between pasted elements, defaults to ""}
}
\description{
Return a vector composed of pasted elements from the input dataframe at the same column.
}
\examples{

print(paste_datf2(inpt_datf=data.frame(c(1, 2, 1), c(33, 22, 55))))

#[1] "121" "332255"

}
