% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{paste_df}
\alias{paste_df}
\title{paste_df}
\usage{
paste_df(inpt_df, sep = "")
}
\arguments{
\item{inpt_df}{is the input dataframe}

\item{sep}{is the separator between pasted elements, defaults to ""}
}
\description{
Return a vector composed of pasted elements from the input dataframe at the same index.
}
\examples{
print(paste_df(inpt_df=data.frame(c(1, 2, 1), c(33, 22, 55))))

[1] "133" "222" "155"
}
