% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{pattern_generator}
\alias{pattern_generator}
\title{pattern_generator}
\usage{
pattern_generator(base_, from_, lngth, hmn = 1, after = 1)
}
\arguments{
\item{base_}{is the pattern that will be kept}

\item{from_}{is the vector from which the element of the varying part will be generated}

\item{hmn}{is how many of varying pattern from the same base will be created}

\item{after}{is set to 1 by default, it means that the varying part will be after the fixed part, set to 0 if you want the varying part to be before}
}
\description{
Allow to create patterns which have a part that is varying aech time randomly
}
