% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{pattern_tuning}
\alias{pattern_tuning}
\title{pattern_tuning}
\usage{
pattern_tuning(pattrn, spe_nb, spe_l, exclude_type, hmn = 1, rg = c(0, 0))

pattern_tuning(pattrn, spe_nb, spe_l, exclude_type, hmn = 1, rg = c(0, 0))
}
\arguments{
\item{pattrn}{is the character that will be tuned}

\item{spe_nb}{is the number of new character that will be replaced}

\item{spe_l}{is the source vector from which the new characters will be replace old ones}

\item{exclude_type}{is character that won't be replaced}

\item{hmn}{is how many output the function will return}

\item{rg}{is a vector with two parameters (index of the first letter that will be replaced, index of the last letter that will be replaced) default is set to all the letters from the source pattern}
}
\description{
Allow to tune a pattern very precisely and output a vector containing its variations n times.

Allow to tune a pattern very precisely
}
