% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{power_to_char}
\alias{power_to_char}
\title{power_to_char}
\usage{
power_to_char(inpt_v = c())
}
\arguments{
\item{inpt_v}{is the input vector containing scientific number, but also other elements that won't be taken in count}
}
\description{
Convert a scientific number to a string representing normally the number.
}
\examples{

print(power_to_char(inpt_v = c(22 * 10000000, 12, 9 * 0.0000002)))

[1] "2200000000" "12"         "0.0000018" 

}
