% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{r_print}
\alias{r_print}
\title{r_print}
\usage{
r_print(inpt_v, sep_ = "and", begn = "This is", end = ", voila!")
}
\arguments{
\item{inpt_v}{is the input vector}

\item{sep_}{is the separator between each elements}

\item{begn}{is the character put at the beginning of the print}

\item{end}{is the character put at the end of the print}
}
\description{
Allow to print vector elements in one row.
}
\examples{

print(r_print(inpt_v=c(1:33)))

#[1] "This is  1 and 2 and 3 and 4 and 5 and 6 and 7 and 8 and 9 and 10 and 11 and 12 and 13 
#and 14 and 15 and 16 and 17 and 18 and 19 and 20 and 21 and 22 and 23 and 24 and 25 and 26 
#and 27 and 28 and 29 and 30 and 31 and 32 and 33 and , voila!"

}
