% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{read_edm_parser}
\alias{read_edm_parser}
\title{read_edm_parser}
\usage{
read_edm_parser(inpt, to_find_v = c())
}
\arguments{
\item{inpt}{is the input dataset}

\item{to_find_v}{is the vector containing the path to find the data, see examples}
}
\description{
Allow to read data from edm parsed dataset, see examples
}
\examples{

print(read_edm_parser("(ok(ee:56))(ok(oui(rr((rr2:6)(rr:5))))(oui(bb(rr2:1)))(ee1:4))", 
to_find_v = c("ok", "oui", "rr", "rr2")))

[1] "6"

print(read_edm_parser("(ok(ee:56))(ok(oui(rr((rr2:6)(rr:5))))(oui(bb(rr2:1)))(ee1:4))", to_find_v = c("ok", "ee")))

[1] "56"

print(read_edm_parser("(ok(ee:56))(ok(oui(rr((rr2:6)(rr:5))))(oui(bb(rr2:1)))(ee1:4))", to_find_v = c("ee")))

[1] "56"

}
