% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{rearangr_v}
\alias{rearangr_v}
\title{rearangr_v}
\usage{
rearangr_v(inpt_v, w_v, how = "increasing")
}
\arguments{
\item{inpt_v}{is the vector that contains the sequance of number}

\item{w_v}{is the vector containing the elements related to inpt_v}

\item{how}{is the way the elements of w_v will be outputed according to if inpt_v will be sorted ascendigly or descendingly}
}
\description{
Reanranges a vector "w_v" according to another vector "inpt_v". inpt_v contains a sequence of number. inpt_v and w_v have the same size and their indexes are related. The output will be a vector containing all the elements of w_v rearanges in descending or asending order according to inpt_v
}
\examples{

print(rearangr_v(inpt_v=c(23, 21, 56), w_v=c("oui", "peut", "non"), how="decreasing"))

#[1] "non"  "oui"  "peut"

}
