% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{regex_spe_detect}
\alias{regex_spe_detect}
\title{regex_spe_detect}
\usage{
regex_spe_detect(inpt)
}
\arguments{
\item{inpt}{the input character}
}
\description{
Takes a character as input and returns its regex-friendly character for R.
}
\examples{

print(regex_spe_detect("o"))

[1] "o"

print(regex_spe_detect("("))

[1] "\\\\("

print(regex_spe_detect("tr(o)m"))

[1] "tr\\\\(o\\\\)m"

print(regex_spe_detect(inpt="fggfg[fggf]fgfg(vg?fgfgf.gf)"))

[1] "fggfg\\\\[fggf\\\\]fgfg\\\\(vg\\\\?fgfgf\\\\.gf\\\\)"

}
