% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{rm_rows}
\alias{rm_rows}
\title{rm_rows}
\usage{
rm_rows(inpt_datf, flagged_vals = c())
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{flagged_vals}{is a vector containing the characters that will drop any rows that contains it}
}
\description{
Allow to remove certain rows that contains certains characters, see examples.
}
\examples{

datf <- data.frame(c(1, 2, NA, 4), c(1:4))
print(datf)

  c.1..2..NA..4. c.1.4.
1              1      1
2              2      2
3             NA      3
4              4      4

print(rm_rows(inpt_datf = datf, flagged_vals = c(1, 4)))

  c.1..2..NA..4. c.1.4.
2              2      2
3             NA      3

}
