% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{row_to_col}
\alias{row_to_col}
\title{row_to_col}
\usage{
row_to_col(inpt_datf)
}
\arguments{
\item{inpt_datf}{is the inout dataframe}
}
\description{
Allow to reverse a dataframe (rows become cols and cols become rows)
}
\examples{

datf_test <- data.frame(c(1, 11), c(2, 10), c(3, 9), c(4, 8))

print(datf_test)

  c.1..11. c.2..10. c.3..9. c.4..8.
1        1        2       3       4
2       11       10       9       8

print(row_to_col(inpt_datf = datf_test))

  1  2
1 1 11
2 2 10
3 3  9
4 4  8

}
