% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{see_df}
\alias{see_df}
\title{see_df}
\usage{
see_df(df, condition_l, val_l, conjunction_l = c(), rt_val = T, f_val = F)
}
\arguments{
\item{df}{is the input dataframe}

\item{condition_l}{is the vector of the possible conditions ("==", ">", "<", "!=", "\%\%") (equal, greater than, lower than, not equal to, is divisible by), you can put the same condition n times.}

\item{val_l}{is the list of vectors containing the values related to condition_l (so the vector of values has to be placed in the same order)}

\item{conjunction_l}{contains the | or & conjunctions, so if the length of condition_l is equal to 3, there will be 2 conjunctions. If the length of conjunction_l is inferior to the length of condition_l minus 1, conjunction_l will match its goal length value with its last argument as the last arguments. For example, c("&", "|", "&") with a goal length value of 5 --> c("&", "|", "&", "&", "&")}

\item{rt_val}{is a special value cell returned when the conditions are respected}

\item{f_val}{is a special value cell returned when the conditions are not respected}
}
\description{
Allow to return a datafame with special value cells (ex: TRUE) where the condition entered are respected and another special value cell (ex: FALSE) where these are not
}
\details{
This function will return an error if number only comparative conditions are given in addition to having character values in the input dataframe.
}
\examples{
see_df(df, c("\%\%", "=="), list(c(2, 11), c(3)), list("|") 
#will return all the values that are divisible by 2 and 11 
#and all the values that are equal to 3 from the dataframe 
}
