% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{see_diff}
\alias{see_diff}
\title{see_diff}
\usage{
see_diff(vec1 = c(), vec2 = c())
}
\arguments{
\item{vec1}{is the first vector}

\item{vec2}{is the second vector}
}
\description{
Output the opposite of intersect(a, b). Already seen at: https://stackoverflow.com/questions/19797954/function-to-find-symmetric-difference-opposite-of-intersection-in-r
}
\examples{

print(see_diff(c(1:7), c(4:12)))

[1] 1 2 3 8 9 10 11 12

}
