% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{see_diff_all}
\alias{see_diff_all}
\title{see_diff_all}
\usage{
see_diff_all(...)
}
\arguments{
\item{...}{are all the input vectors}
}
\description{
Allow to perform the opposite of intersect function to n vectors.
}
\examples{

vec1 <- c(3:6)
vec2 <- c(1:8)
vec3 <- c(12:16)

print(see_diff_all(vec1, vec2))

[1] 1 2 7 8

print(see_diff_all(vec1, vec2, vec3))

[1]  3  4  5  6  1  2  7  8 12 13 14 15 16

}
