% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{see_diff_detailled}
\alias{see_diff_detailled}
\title{see_diff_detailled}
\usage{
see_diff_detailled(vec1 = c(), vec2 = c())
}
\arguments{
\item{vec1}{is one of the input vector}

\item{vec2}{is the other input vector}
}
\description{
Behaves exactly like the see_diff function but is written more explicitely, see examples
}
\examples{

print(see_diff_detailled(c(1:6), c(3:9)))

[1] 1 2 7 8 9

}
