% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{see_in_grep}
\alias{see_in_grep}
\title{see_in_grep}
\usage{
see_in_grep(from_v = c(), in_v = c())
}
\arguments{
\item{from_v}{is the vector that may contains elements that contains the same patterns that those in in_v, see examples}

\item{in_v}{is a vector that contains the patterns to find}
}
\description{
Allow to get the indices of the elements of a vector that contains certyain patterns.
The type of the output may change in function of the input vectors, see examples
}
\examples{

print(see_in_grep(from_v = c("oui", "non", "peut"), 
                  in_v = c("ou", "eu", "plm")))

           ou            eu           plm 
           1             3            -1 

 print(see_in_grep(from_v = c("oui", "non", "peut", "oui"), 
      in_v = c("ou", "eu", "plm")))

$ou
[1] 1 4

$eu
[1] 3

$plm
[1] -1

}
