% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{see_in_l}
\alias{see_in_l}
\title{see_in_l}
\usage{
see_in_l(from_v = c(), in_v = c())
}
\arguments{
\item{from_v}{is the vector that may contains elements that contains the same patterns that those in in_v, see examples}

\item{in_v}{is a vector that contains the patterns to find}
}
\description{
Allow to get the patterns that are present in the elements of a vector, see examples
}
\examples{

print(see_in_l(from_v = c("oui", "non", "peut"), 
  in_v = c("ou", "pe", "plm")))

   ou    pe   plm 
  TRUE  TRUE FALSE  

}
