% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{see_mode}
\alias{see_mode}
\title{see_mode}
\usage{
see_mode(inpt_v = c())
}
\arguments{
\item{inpt_v}{is the input vector}
}
\description{
Allow to get the mode of a vector, see examples.
}
\examples{

print(see_mode(inpt_v = c(1, 1, 2, 2, 2, 3, 1, 2)))

[1] 2

print(see_mode(inpt_v = c(1, 1, 2, 2, 2, 3, 1)))

[1] 1

}
