% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{selected_char}
\alias{selected_char}
\title{selected_char}
\usage{
selected_char(n, base_char = letters)
}
\arguments{
\item{n}{is how many sequence of numbers will be generated}

\item{base_char}{is the vector containing the elements from which the character is generated}
}
\description{
Allow to generate a char based on a conbinaison on characters from a vector and a number
}
\examples{

print(selected_char(1222))

[1] "zta"

}
