% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{sequence_na_mean1}
\alias{sequence_na_mean1}
\title{sequence_na_mean1}
\usage{
sequence_na_mean1(inpt_datf, bf_, step = 1)
}
\arguments{
\item{inpt_datf}{is the input dataframe}
}
\description{
In a dataframe generated by the function historic_sequence, convert all NA to the mean of the values at the same variable for the individual at the id where the NA occurs, see examples (only accepts numeric variables)
}
\examples{

set.seed(123)
var1 <- round(runif(n = 14, min = 100, max = 122))
set.seed(123)
var2 <- round(runif(n = 14, min = 14, max = 20))

datf <- data.frame("ids" = c(20, 20, 20, 20, 19, 19, 19, 18, 18, 18, 18,
17, 17, 17),
"individual" = c("oui", "non", "peut1", "peut2",
"oui", "peut1", "peut2"),
"var1" = var1,
"var2" = var2)
datf <- historic_sequence(inpt_datf = datf, bf_ = 2)
datf[3, 4] <- NA
datf[6, 4] <- NA
datf[1, 3] <- NA
print(datf)

  id_seq individual var1-1 var1-2 var2-1 var2-2
1     20        oui     NA    120     20     19
2     20        non     NA    112     NA     17
3     20      peut1    101     NA     14     17
4     20      peut2    112    121     17     20
5     19        oui    120    110     19     17
6     19      peut1    110     NA     17     18
7     19      peut2    121    113     20     17

print(sequence_na_mean1(inpt_datf = datf, bf_ = 2))

  id_seq individual var1-1 var1-2 var2-1 var2-2
1     20        oui    115  120.0     20     19
2     20        non    112  112.0     17     17
3     20      peut1    101  105.5     14     17
4     20      peut2    112  121.0     17     20
5     19        oui    120  110.0     19     17
6     19      peut1    110  105.5     17     18
7     19      peut2    121  113.0     20     17 

}
