% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{sequence_na_med2}
\alias{sequence_na_med2}
\title{sequence_na_med2}
\usage{
sequence_na_med2(inpt_datf, bf_)
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{bf_}{is how at how many n -1 we look for the value of the variables for the individual at time index n}
}
\description{
In a dataframe generated by the function historic_sequence2, convert all NA to the median of the values at the same variable for the individual at the id where the NA occurs, see examples (only accepts numeric variables)
}
\examples{

set.seed(123)
var1 <- round(runif(n = 14, min = 100, max = 122))
set.seed(123)
var2 <- round(runif(n = 14, min = 14, max = 20))
datf <- data.frame("ids" = c(20, 20, 20, 20, 19, 19, 19, 18, 18, 18, 18,
17, 17, 17),
"individual" = c("oui", "non", "peut1", "peut2",
"oui", "peut1", "peut2"),
"var1" = var1,
"var2" = var2)
datf <- historic_sequence2(inpt_datf = datf, bf_ = 2)
datf[3, 4] <- NA
datf[6, 4] <- NA
datf[1, 3] <- NA
print(datf)

  id_seq individual var1-0 var1-1 var1-2 var2-0 var2-1 var2-2
1     20        oui     NA    121    120     16     20     19
2     20        non    117     NA    112     19     NA     17
3     20      peut1    109     NA    110     16     14     17
4     20      peut2    119    112    121     19     17     20
5     19        oui    121    120    110     20     19     17
6     19      peut1    101     NA    115     14     17     18
7     19      peut2    112    121    113     17     20     17

print(sequence_na_med2(inpt_datf = datf, bf_ = 2))

  id_seq individual var1-0 var1-1 var1-2 var2-0 var2-1 var2-2
1     20        oui    120  121.0    120     16     20     19
2     20        non    117  114.5    112     19     18     17
3     20      peut1    109  109.0    110     16     14     17
4     20      peut2    119  112.0    121     19     17     20
5     19        oui    121  120.0    110     20     19     17
6     19      peut1    101  109.0    115     14     17     18
7     19      peut2    112  121.0    113     17     20     17

}
