% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{sort_date}
\alias{sort_date}
\title{sort_date}
\usage{
sort_date(inpt_v, frmt, sep_ = "-", ascending = FALSE, give = "value")
}
\arguments{
\item{inpt_v}{is the input vector containing all the dates}

\item{frmt}{is the format  of the dates, (any combinaison of letters "s" for second, "n", for minute, "h" for hour, "d" for day, "m" for month and "y" for year)}

\item{sep_}{is the separator used for the dates}

\item{ascending}{is the used to sort the dates}

\item{give}{takes only two values "index" or "value", if give == "index", the function will output the index of sorted dates from inpt_v, if give == "value", the function will output the value, it means directly the sorted dates in inpt_v, see examples}
}
\description{
Allow to sort any vector containing a date, from any kind of format (my, hdmy, ymd ...), see examples.
}
\examples{

print(sort_date(inpt_v = c("01-11-2025", "08-08-1922", "12-04-1966")
                , frmt = "dmy", sep_ = "-", ascending = TRUE, give = "value"))

[1] "08-08-1922" "12-04-1966" "01-11-2025"

print(sort_date(inpt_v = c("01-11-2025", "08-08-1922", "12-04-1966")
                , frmt = "dmy", sep_ = "-", ascending = FALSE, give = "value"))

[1] "01-11-2025" "12-04-1966" "08-08-1922"

print(sort_date(inpt_v = c("01-11-2025", "08-08-1922", "12-04-1966")
                , frmt = "dmy", sep_ = "-", ascending = TRUE, give = "index"))

[1] 2 3 1

print(sort_date(inpt_v = c("22-01-11-2025", "11-12-04-1966", "12-12-04-1966")
                , frmt = "hdmy", sep_ = "-", ascending = FALSE, give = "value"))

[1] "22-01-11-2025" "12-12-04-1966" "11-12-04-1966"

print(sort_date(inpt_v = c("03-22-01-11-2025", "56-11-12-04-1966", "23-12-12-04-1966")
                , frmt = "nhdmy", sep_ = "-", ascending = FALSE, give = "value"))

[1] "03-22-01-11-2025" "23-12-12-04-1966" "56-11-12-04-1966"

}
