% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{split_by_step}
\alias{split_by_step}
\title{split_by_step}
\usage{
split_by_step(inpt_v, by)
}
\arguments{
\item{inpt_v}{is the input character or vector of characters}

\item{by}{is the step}
}
\description{
Allow to split a string or a vector of strings by a step, see examples.
}
\examples{

print(split_by_step(inpt_v = c("o", "u", "i", "n", "o", "o", "u", "i", "o", "Z"), by = 2))

[1] "ou" "in" "oo" "ui" "oZ"

print(split_by_step(inpt_v = c("o", "u", "i", "n", "o", "o", "u", "i", "o", "Z"), by = 3))

[1] "oui" "noo" "uio" "Z"  

print(split_by_step(inpt_v = c("o", "u", "i", "n", "o", "o", "u", "i", "o", "Z"), by = 4))

[1] "ouin" "ooui" "oZ"  

print(split_by_step(inpt_v = 'ouinoouioz', by = 4))

[1] "ouin" "ooui" "oZ"   

}
