% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{successive_diff}
\alias{successive_diff}
\title{successive_diff}
\usage{
successive_diff(inpt_v)
}
\arguments{
\item{inpt_v}{is the input numeric vector}
}
\description{
Allow to see the difference beteen the suxxessive elements of an numeric vector
}
\examples{

print(successive_diff(c(1:10)))

[1] 1 1 1 1 1

print(successive_diff(c(1:11, 13, 19)))

[1] 1 1 1 1 1 2 6

}
