% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{sum_group2}
\alias{sum_group2}
\title{sum_group2}
\usage{
sum_group2(inpt_datf, col_grp = c(), col_to_add = c())
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{col_grp}{is a vector containing the column names or the column numbers of the groups}

\item{col_to_add}{is a vector containing the column names or the column numbers of the variables to aggregate}
}
\description{
Allow to aggregate variables according to groups, see examples
}
\examples{

set.seed(123)
datf <- data.frame("country" = c("France", "Germany", "France", "Italy", "Italy", "France"),
                   "year" = c(2012, 2012, 2013, 2011, 2012, 2011),
                   "comp_arm" = c("higher", "lower", "higher", "higher", "lower", "lower"),
                   "pop" = runif(n = 6, min = 65000000, max = 69000000),
                   "random_var" = round(x = runif(n = 6, min = 16, max = 78), digits = 0))
datf

  country year comp_arm      pop random_var
1  France 2012   higher 66150310         49
2 Germany 2012    lower 68153221         71
3  France 2013   higher 66635908         50
4   Italy 2011   higher 68532070         44
5   Italy 2012    lower 68761869         75
6  France 2011    lower 65182226         44

print(sum_group2(inpt_datf = datf, col_grp = c("country"), col_to_add = c("random_var", "pop")))

  country year comp_arm       pop random_var
1  France 2012   higher 197968444        143
3  France 2013   higher 197968444        143
6  France 2011    lower 197968444        143
2 Germany 2012    lower  68153221         71
4   Italy 2011   higher 137293939        119
5   Italy 2012    lower 137293939        119

print(sum_group2(inpt_datf = datf, col_grp = c("year"), col_to_add = c("random_var", "pop")))

  country year comp_arm       pop random_var
1  France 2012   higher 203065400        195
2 Germany 2012    lower 203065400        195
5   Italy 2012    lower 203065400        195
3  France 2013   higher  66635908         50
4   Italy 2011   higher 133714296         88
6  France 2011    lower 133714296         88

print(sum_group2(inpt_datf = datf, col_grp = c("country", "year"), col_to_add = c("random_var")))

  country year comp_arm      pop random_var
1  France 2012   higher 66150310         49
2 Germany 2012    lower 68153221         71
3  France 2013   higher 66635908         50
4   Italy 2011   higher 68532070         44
5   Italy 2012    lower 68761869         75
6  France 2011    lower 65182226         44

set.seed(123)
pop_v <- runif(n = 6, min = 65000000, max = 69000000)
pop_v[c(1, 3)] <- NA
set.seed(123)
datf <- data.frame("country" = c("France", "Germany", "France", "Italy", "Italy", "France"),
                   "year" = c(2012, 2012, 2013, 2011, 2012, 2011),
                   "comp_arm" = c("higher", "lower", "higher", "higher", "lower", "lower"),
                   "pop" = pop_v,
                   "random_var" = round(x = runif(n = 6, min = 16, max = 78), digits = 0))
datf
  country year comp_arm      pop random_var
1  France 2012   higher       NA         34
2 Germany 2012    lower 68153221         65
3  France 2013   higher       NA         41
4   Italy 2011   higher 68532070         71
5   Italy 2012    lower 68761869         74
6  France 2011    lower 65182226         19

print(sum_group2(inpt_datf = datf, col_grp = c("year"), col_to_add = c("random_var", "pop")))

  country year comp_arm       pop random_var
1  France 2012   higher 136915090        173
2 Germany 2012    lower 136915090        173
5   Italy 2012    lower 136915090        173
3  France 2013   higher        NA         41
4   Italy 2011   higher 133714296         90
6  France 2011    lower 133714296         90

}
