% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{swipr}
\alias{swipr}
\title{swipr}
\usage{
swipr(inpt_datf, how_to = c(), id_w = 2, id_ids = 1)
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{how_to}{is a vector containing the elements in the order wanted}

\item{id_w}{is the column number or the column name of the elements}

\item{id_ids}{is the column number or the column name of the ids}
}
\description{
Returns an ordered dataframes according to the elements order given. The input datafram has two columns, one with the ids which can be bonded to multiple elements in the other column.
}
\examples{

datf <- data.frame("col1"=c("Af", "Al", "Al", "Al", "Arg", "Arg", "Arg", "Arm", "Arm", "Al"),

        "col2"=c("B", "B", "G", "S", "B", "S", "G", "B", "G", "B"))

print(swipr(inpt_datf=datf, how_to=c("G", "S", "B")))

   col1 col2
1    Af    B
2    Al    G
3    Al    S
4    Al    B
5   Arg    G
6   Arg    S
7   Arg    B
8   Arm    G
9   Arm    B
10   Al    B

}
