% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{time_serie_equalizer}
\alias{time_serie_equalizer}
\title{time_serie_equalizer}
\usage{
time_serie_equalizer(
  inpt_datf,
  time_col,
  null_value = 0,
  individual_col,
  var_col = c()
)
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{time_col}{is the column number or name of the time values}

\item{null_value}{is the default value of the variable of the elements that will be added}

\item{individual_col}{is the column name or number of the individuals}

\item{var_col}{is a vector containing the column names or numbers if the variables that will equal the null_value for the individual at the new time values}
}
\description{
Allow ewualize the occurence of each elements in all the timestamps, see examples
}
\examples{

print(datf)

   individual country year                           energy_source twh_cons
1           A  France 1995                     biofuel_electricity     1.82
2           A  France 1996                        coal_electricity    24.18
3           A  France 1997                         gas_electricity     3.84
4           A  France 1998                       hydro_electricity    71.33
5           A  France 1999                     nuclear_electricity   377.23
6           A  France 2000                         oil_electricity    10.50
7           A  France 2001 other_renewable_exc_biofuel_electricity     0.51
10          B  France 1995                     biofuel_electricity     9.50
11          B  France 1996                        coal_electricity     2.16
12          B  France 1997                         gas_electricity    31.43
13          B  France 1998                       hydro_electricity    53.19
14          B  France 1999                     nuclear_electricity   335.65
15          B  France 2000                         oil_electricity     9.71
16          B  France 2001 other_renewable_exc_biofuel_electricity     0.60
17          B  France 2002                       solar_electricity    23.26
18          B  France 2003                        wind_electricity    48.61

print(time_serie_equalizer(inpt_datf = datf,
                           time_col = "year",
                           null_value = 0,
                           individual_col = 1,
                           var_col = "twh_cons"))

    individual country year                           energy_source twh_cons
1            A  France 1995                     biofuel_electricity     1.82
2            A  France 1996                        coal_electricity    24.18
3            A  France 1997                         gas_electricity     3.84
4            A  France 1998                       hydro_electricity    71.33
5            A  France 1999                     nuclear_electricity   377.23
6            A  France 2000                         oil_electricity    10.50
7            A  France 2001 other_renewable_exc_biofuel_electricity     0.51
8            B  France 1995                     biofuel_electricity     9.50
9            B  France 1996                        coal_electricity     2.16
10           B  France 1997                         gas_electricity    31.43
11           B  France 1998                       hydro_electricity    53.19
12           B  France 1999                     nuclear_electricity   335.65
13           B  France 2000                         oil_electricity     9.71
14           B  France 2001 other_renewable_exc_biofuel_electricity     0.60
15           B  France 2002                       solar_electricity    23.26
16           B  France 2003                        wind_electricity    48.61
17           A  France 2002                     biofuel_electricity     0.00
18           A  France 2003                     biofuel_electricity     0.00

}
