% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{to_unique}
\alias{to_unique}
\title{to_unique}
\usage{
to_unique(inpt_v, distinct_type = "suffix", distinct_val = "number", sep = "-")
}
\arguments{
\item{inpt_v}{is the input vectors}

\item{distinct_type}{takes two values: suffix or prefix}

\item{distinct_val}{takes two values: number (unique sequence of number to differencfiate each value) or letter (unique sequence of letters to differenciate each value)}
}
\description{
Allow to transform a vector containing elements that have more than 1 occurence to a vector with only uniques elements.
}
\examples{

print(to_unique(inpt_v = c("a", "a", "e", "a", "i", "i"), 
                distinct_type = "suffix", 
                distinct_val = "number", 
                sep = "-"))

[1] "a-1" "a-2" "e"   "a-3" "i-1" "i-2"

print(to_unique(inpt_v = c("a", "a", "e", "a", "i", "i"), 
                distinct_type = "suffix", 
                distinct_val = "letter", 
                sep = "-"))

[1] "a-a" "a-b" "e"   "a-c" "i-a" "i-b"

print(to_unique(inpt_v = c("a", "a", "e", "a", "i", "i"), 
                distinct_type = "prefix", 
                distinct_val = "number", 
                sep = "/"))

[1] "1/a" "2/a" "e"   "3/a" "1/i" "2/i"

print(to_unique(inpt_v = c("a", "a", "e", "a", "i", "i"), 
                distinct_type = "prefix", 
                distinct_val = "letter", 
                sep = "_"))

[1] "a_a" "b_a" "e"   "c_a" "a_i" "b_i" 

}
