% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{union_all}
\alias{union_all}
\title{union_all}
\usage{
union_all(...)
}
\arguments{
\item{...}{are all the input vectors}
}
\description{
Allow to perform a union function to n vectors.
}
\examples{

print(union_all(c(1, 2), c(3, 4), c(1:8)))

[1] 1 2 3 4 5 6 7 8

print(union_all(c(1, 2), c(3, 4), c(7:8)))

[1] 1 2 3 4 7 8

}
