% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{union_keep}
\alias{union_keep}
\title{union_keep}
\usage{
union_keep(...)
}
\arguments{
\item{...}{are all the input vectors}
}
\description{
Performs a union operation keeping the number of elements of all input vectors, see examples
}
\examples{

print(union_keep(c("a", "ee", "ee"), c("p", "p", "a", "i"), c("a", "a", "z")))

[1] "a"  "ee" "ee" "p"  "p"  "i"  "z" 

print(union_keep(c("a", "ee", "ee"), c("p", "p", "a", "i")))

[1] "a"  "ee" "ee" "p"  "p"  "i" 

}
