% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{unique_df}
\alias{unique_df}
\title{unique_df}
\usage{
unique_df(inpt_df, col = F)
}
\arguments{
\item{inpt_df}{is the input dataframe}

\item{col}{is a parameter that specifies if the dataframe returned should have unique columns or rows, defaults to F, so the dataframe returned by default has unique rows}
}
\description{
Returns the input dataframe with the unique columns or rows.
}
\examples{

df1 <- data.frame(c(1, 2, 1, 3), c("a", "z", "a", "p"))

print(unique_df(inpt_df=df1))

   c.1..2..1..3. c..a....z....a....p..
1             1                     a
2             2                     z
4             3                     p

df1 <- data.frame(c(1, 2, 1, 3), c("a", "z", "a", "p"), c(1, 2, 1, 3))

print(unique_df(inpt_df=df1, col=T))

  cur_v cur_v
1     1     a
2     2     z
3     1     a
4     3     p

}
