% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{unique_ltr_from_v}
\alias{unique_ltr_from_v}
\title{unique_ltr_from_v}
\usage{
unique_ltr_from_v(inpt_v, keep_v = c("?", "!", ":", "&", ",", ".", letters))
}
\arguments{
\item{inpt_v}{is the input vector containing all the elements}

\item{keep_v}{is the vector containing all the characters that the elements in inpt_v may contain}
}
\description{
Returns the unique characters contained in all the elements from an input vector "inpt_v"
}
\examples{

print(unique_ltr_from_v(inpt_v=c("bonjour", "lpoerc", "nonnour", "bonnour", "nonjour", "aurevoir")))

#[1] "b" "o" "n" "j" "u" "r" "l" "p" "e" "c" "a" "v" "i" 

}
