% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{unique_pos}
\alias{unique_pos}
\title{unique_pos}
\usage{
unique_pos(vec)
}
\arguments{
\item{vec}{is the input vector}
}
\description{
Allow to find indexes of the unique values from a vector.
}
\examples{
print(unique_pos(c(3, 4, 3, 5, 6)))
[1] 1 2 4 5
}
