% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{unique_total}
\alias{unique_total}
\title{unique_total}
\usage{
unique_total(inpt_v = c())
}
\arguments{
\item{inpt_v}{is the input vector containing all the elements}
}
\description{
Returns a vector with the total amount of occurences for each element in the input vector. The occurences of each element follow the same order as the unique function does, see examples
}
\examples{

print(unique_total(inpt_v = c(1:12, 1)))

 [1] 2 1 1 1 1 1 1 1 1 1 1 1

print(unique_total(inpt_v = c(1:12, 1, 11, 11)))

 [1] 2 1 1 1 1 1 1 1 1 1 3 1

}
