% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{v_to_df}
\alias{v_to_df}
\title{v_to_df}
\usage{
v_to_df(inpt_v, sep = "-")
}
\arguments{
\item{inpt_v}{is the input vector}

\item{sep}{is the separator used to seprate the columns}
}
\description{
Allow to convert a vector to a dataframe according to a separator.
}
\examples{
v <- c("aa-yy-uu", "zz-gg-hhh", "zz-gg-hhh", "zz-gg-hhh")
v_to_df(inpt_v=v, sep="-")
}
