% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{vec_in_df}
\alias{vec_in_df}
\title{vec_in_df}
\usage{
vec_in_df(inpt_df, inpt_vec = c(), coeff = 0, stop_untl = 1, conventional = F)
}
\arguments{
\item{inpt_df}{is the input dataframe}

\item{inpt_vec}{is the vector that may be in the input dataframe}

\item{coeff}{is the "slope coefficient" of inpt_vec}

\item{stop_untl}{is the maximum number of the input vector the function returns, if in the dataframe}

\item{conventinal}{is if a positive slope coefficient means that the vector goes upward or downward}
}
\description{
Allow to get if a vector is in a dataframe. Returns the row and column of the vector in the dataframe if the vector is contained in the dataframe.
}
\examples{

df1 <- data.frame(c(1:5), c(5:1), c("a", "z", "z", "z", "a"))

print(df1)

  c.1.5. c.5.1. c..a....z....z....z....a..
1      1      5                          a
2      2      4                          z
3      3      3                          z
4      4      2                          z
5      5      1                          a

print(vec_in_df(inpt_df=df1, inpt_vec=c(5, 4, "z"), coeff=1))

NULL

print(vec_in_df(inpt_df=df1, inpt_vec=c(5, 2, "z"), coeff=1))

[1] 5 1

print(vec_in_df(inpt_df=df1, inpt_vec=c(3, "z"), coeff=1))

[1] 3 2

print(vec_in_df(inpt_df=df1, inpt_vec=c(4, "z"), coeff=-1))

[1] 2 2

print(vec_in_df(inpt_df=df1, inpt_vec=c(2, 3, "z"), coeff=-1))

[1] 2 1

print(vec_in_df(inpt_df=df1, inpt_vec=c(5, 2, "z"), coeff=-1, conventional=T))
 
[1] 5 1

df1[4, 2] <- 1

print(vec_in_df(inpt_df=df1, inpt_vec=c(1, "z"), coeff=-1, conventional=T, stop_untl=4))

[1] 4 2 5 2

}
