% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{vector_replacor}
\alias{vector_replacor}
\title{vector_replacor}
\usage{
vector_replacor(inpt_v = c(), sus_val = c(), rpl_val = c(), grep_ = FALSE)
}
\arguments{
\item{inpt_v}{is the input vector}

\item{sus_val}{is a vector containing all the values that will be replaced}

\item{rpl_val}{is a vector containing the value of the elements to be replaced (sus_val), so sus_val and rpl_val should be the same size}

\item{grep_}{is if the elements in sus_val should be equal to the elements to replace in inpt_v or if they just should found in the elements}
}
\description{
Allow to replace certain values in a vector.
}
\examples{

print(vector_replacor(inpt_v=c(1:15), sus_val=c(3, 6, 8, 12), 
     rpl_val=c("oui", "non", "e", "a")))

# [1] "1"   "2"   "oui" "4"   "5"   "non" "7"   "e"   "9"   "10"  "11"  "a"  
#[13] "13"  "14"  "15" 

print(vector_replacor(inpt_v=c("non", "zez", "pp a ftf", "fdatfd", "assistance", 
"ert", "repas", "repos"), 
sus_val=c("pp", "as", "re"), rpl_val=c("oui", "non", "zz"), grep_=TRUE))

#[1] "non"  "zez"  "oui"  "fdatfd" "non"  "ert"  "non"  "zz"  

}
