% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{vlookup_df}
\alias{vlookup_df}
\title{vlookup_df}
\usage{
vlookup_df(df, v_id, col_id = 1, included_col_id = "yes")
}
\arguments{
\item{df}{is the input dataframe}

\item{v_id}{is a vector containing the ids}

\item{col_id}{is the column that contains the ids (default is equal to 1)}

\item{included_col_id}{is if the result should return the col_id (default set to yes)}
}
\description{
Alow to perform a vlookup on a dataframe
}
