% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{wider_df}
\alias{wider_df}
\title{wider_df}
\usage{
wider_df(inpt_df, col_to_splt = c(), sep_ = "-")
}
\arguments{
\item{inpt_df}{is the input dataframe}

\item{col_to_splt}{is a vector containing the number or the colnames of the columns to split according to a separator}

\item{sep_}{is the separator of the elements to split to new columns in the input dataframe}
}
\description{
Takes a dataframe as an input and the column to split according to a seprator.
}
\examples{
df1 <- data.frame(c(1:5), c("o-y", "hj-yy", "er-y", "k-ll", "ooo-mm"), c(5:1))

df2 <- data.frame(c(1:5), c("o-y", "hj-yy", "er-y", "k-ll", "ooo-mm"))
 
print(wider_df(inpt_df=df1, col_to_splt=c(2), sep_="-"))

       pre_df X.o.  X.y.  
o-y    1      "o"   "y"  5
hj-yy  2      "hj"  "yy" 4
er-y   3      "er"  "y"  3
k-ll   4      "k"   "ll" 2
ooo-mm 5      "ooo" "mm" 1

print(wider_df(inpt_df=df2, col_to_splt=c(2), sep_="-"))

       pre_df X.o.  X.y.
o-y    1      "o"   "y" 
hj-yy  2      "hj"  "yy"
er-y   3      "er"  "y" 
k-ll   4      "k"   "ll"
ooo-mm 5      "ooo" "mm"
}
