% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{write_edm_parser}
\alias{write_edm_parser}
\title{write_edm_parser}
\usage{
write_edm_parser(inpt, to_write_v, write_data)
}
\arguments{
\item{inpt}{is the input dataset}

\item{to_write_v}{is the vector containing the path to write the data, see examples}
}
\description{
Allow to write data to edm parsed dataset, see examples
}
\examples{

print(write_edm_parser("(ok(ee:56))(ok(oui(rr((rr2:6)(rr:5))))(oui(bb(rr2:1)))(ee1:4))", 
to_write_v = c("ok", "ee"), write_data = c("ii", "olm")))

[1] "(ok(ee:56)(ii:olm))(ok(oui(rr((rr2:6)(rr:5))))(oui(bb(rr2:1)))(ee1:4))"

print(write_edm_parser("(ok(ee:56))(ok(oui(rr((rr2:6)(rr:5))))(oui(bb(rr2:1)))(ee1:4))", 
to_write_v = c("ok", "oui"), write_data = c("ii", "olm")))

[1] "(ok(ee:56))(ok(oui(rr((rr2:6)(rr:5))))(ii:olm)(oui(bb(rr2:1)))(ee1:4))"

print(write_edm_parser("(ok(ee:56))(ok(oui(rr((rr2:6)(rr:5))))(oui(bb(rr2:1)))(ee1:4))", 
to_write_v = c("ok", "oui", "oui"), write_data = c("ii", "olm")))

[1] "(ok(ee:56))(ok(oui(rr((rr2:6)(rr:5))))(oui(bb(rr2:1)))(ii:olm)(ee1:4))"

print(write_edm_parser("", 
to_write_v = c(), write_data = c("ii", "olm")))

[1] "(ii:olm)"

}
