% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{all_stat}
\alias{all_stat}
\title{all_stat}
\usage{
all_stat(inpt_v, var_add = c(), stat_var = c(), inpt_df)
}
\description{
Allow to see all the main statistics indicators (mean, median, variance, standard deviation, sum) of variables in a dataframe by the modality of a variable in a column of the input datarame. In addition to that, you can get the occurence of other qualitative variables by your chosen qualitative variable, you have just to precise it in the vector "stat_var" where all the statistics indicators are given with "occu-var_you_want/".
}
\examples{
df <- data.frame("mod"=c("first", "seco", "seco", "first", "first", "third", "first"), 
                "var1"=c(11, 22, 21, 22, 22, 11, 9), 
               "var2"=c("d", "d", "z", "z", "z", "d", "z"), 
               "var3"=c(45, 44, 43, 46, 45, 45, 42),
              "var4"=c("A", "A", "A", "A", "B", "C", "C"))

all_stat(inpt_v=c("first", "seco"), var_add = c("var1", "var2", "var3", "var4"), 
 stat_var=c("sum", "mean", "median", "sd", "occu-var2/", "occu-var4/", "variance"), 
 inpt_df=df)

  modal_v var_vector occu sum mean  med standard_devaition         variance
  first                                                                  
              var1       64   16 16.5   6.97614984548545 48.6666666666667
            var2-d    1                                                  
            var2-z    3                                                  
              var3      178 44.5   45   1.73205080756888                3
            var4-A    2                                                  
            var4-B    1                                                  
            var4-C    1                                                  
   seco                                                                  
              var1       43 21.5 21.5  0.707106781186548              0.5
            var2-d    1                                                  
            var2-z    1                                                  
              var3       87 43.5 43.5  0.707106781186548              0.5
            var4-A    2                                                  
            var4-B    0                                                  
            var4-C    0                                                  
}
