% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{all_stat}
\alias{all_stat}
\title{all_stat}
\usage{
all_stat(inpt_v, var_add = c(), stat_var = c(), inpt_datf)
}
\arguments{
\item{inpt_v}{is the modalities of the variables}

\item{var_add}{is the variables you want to get the stats from}

\item{stat_var}{is the stats indicators you want}

\item{inpt_datf}{is the input dataframe}
}
\description{
Allow to see all the main statistics indicators (mean, median, variance, standard deviation, sum, max, min, quantile) of variables in a dataframe by the modality of a variable in a column of the input datarame. In addition to that, you can get the occurence of other qualitative variables by your chosen qualitative variable, you have just to precise it in the vector "stat_var" where all the statistics indicators are given with "occu-var_you_want/".
}
\examples{

datf <- data.frame("mod"=c("first", "seco", "seco", "first", "first", "third", "first"), 
                "var1"=c(11, 22, 21, 22, 22, 11, 9), 
               "var2"=c("d", "d", "z", "z", "z", "d", "z"), 
               "var3"=c(45, 44, 43, 46, 45, 45, 42),
              "var4"=c("A", "A", "A", "A", "B", "C", "C"))

print(all_stat(inpt_v=c("first", "seco"), var_add = c("var1", "var2", "var3", "var4"), 
 stat_var=c("sum", "mean", "median", "sd", "occu-var2/", "occu-var4/", "variance", 
"quantile-0.75/"), 
 inpt_datf=datf))

#   modal_v var_vector occu sum mean  med standard_devaition         variance
#1    first                                                                  
#2                var1       64   16 16.5   6.97614984548545 48.6666666666667
#3              var2-d    1                                                  
#4              var2-z    3                                                  
#5                var3      178 44.5   45   1.73205080756888                3
#6              var4-A    2                                                  
#7              var4-B    1                                                  
#8              var4-C    1                                                  
#9     seco                                                                  
#10               var1       43 21.5 21.5  0.707106781186548              0.5
#11             var2-d    1                                                  
#12             var2-z    1                                                  
#13               var3       87 43.5 43.5  0.707106781186548              0.5
#14             var4-A    2                                                  
#15             var4-B    0                                                  
#16             var4-C    0                                                  
#   quantile-0.75
#1               
#2             22
#3               
#4               
#5          45.25
#6               
#7               
#8               
#9               
#10         21.75
#11              
#12              
#13         43.75
#14              
#15              
#16              

}
