% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{edm_pivot_longer2}
\alias{edm_pivot_longer2}
\title{edm_pivot_longer2}
\usage{
edm_pivot_longer2(inpt_datf, col_vars = c(), col_vars_to = c())
}
\arguments{
\item{inpt_datf}{is the input dataframe}

\item{col_vars}{is a vector containing the column names or column numbers of the variables}

\item{col_vars_to}{is a vector containing the varaiables to which will be assign the modalities, see examples}

\item{individual_col}{is the column name or the column number of the individuals}
}
\description{
Performs a pivot longer on dataframe keeping the null values, see examples. The synthax for variables must be value_id-modalitie_var1.modalitie_var2...
}
\examples{

datf <- data.frame("individuals" = c(1, 2, 3),
                   c(1, 2, 3),
                   c(6, 0, 2),
                   c(7, 0, 0),
                   c(0, 0, 0),
                   c(1, 0, 4),
                   c(3, 0, 8),
                   c(9, 0 , 0),
                   c(11, 0, 5))

colnames(datf)[2:ncol(datf)] <- c("val1-A.R", 
                                  "val1-A.T", 
                                  "val1-B.R",
                                  "val1-B.T", 
                                  "val2-A.R",
                                  "val2-A.T",
                                  "val2-B.R",
                                  "val2-B.T")

datf2 <- data.frame("individuals" = c(1, 2, 3),
                   c(7, 0, 2),
                   c(1, 0, 4),
                   c(9, 0, 8),
                   c(11, 22, 5))
colnames(datf2)[2:ncol(datf2)] <- c(

                        "val1-A", 
                        "val1-B",
                        "val2-A",
                        "val2-B"
                   )

print(datf)

  individuals val1-A.R val1-A.T val1-B.R val1-B.T val2-A.R val2-A.T val2-B.R
1           1        1        6        7        0        1        3        9
2           2        2        0        0        0        0        0        0
3           3        3        2        0        0        4        8        0
  val2-B.T
1       11
2        0
3        5

print(edm_pivot_longer2(inpt_datf = datf, 
                          col_vars = c(2:9), 
                          individual_col = 1, 
                          col_vars_to = c("Shape", "Way")))

   individuals Shape Way val1 val2
1            1     A   R    1    1
2            1     A   T    6    3
3            1     B   R    7    9
4            1     B   T    0   11
5            2     A   R    2    0
6            2     A   T    0    0
7            2     B   R    0    0
8            2     B   T    0    0
9            3     A   R    3    4
10           3     A   T    2    8
11           3     B   R    0    0
12           3     B   T    0    5

print(datf2)

  individuals val1-A val1-B val2-A val2-B
1           1      7      1      9     11
2           2      0      0      0     22
3           3      2      4      8      5

print(edm_pivot_longer2(inpt_datf = datf2, 
                        col_vars = c(2:5), 
                        individual_col = 1, 
                        col_vars_to = c("Shape")))

  individuals Shape val1 val2
1           1     A    7    9
2           1     B    1   11
3           2     A    0    0
4           2     B    0   22
5           3     A    2    8
6           3     B    4    5



print(cur_data)

      individual country year twh_cons-biofuel_electricity
7475 France_1995  France 1995                         1.82
7503 France_2023  France 2023                         9.50
     twh_cons-coal_electricity twh_cons-gas_electricity
7475                     24.18                     3.84
7503                      2.16                    31.43
     twh_cons-hydro_electricity twh_cons-nuclear_electricity
7475                      71.33                       377.23
7503                      53.19                       335.65
     twh_cons-oil_electricity twh_cons-other_renewable_exc_biofuel_electricity
7475                    10.50                                             0.51
7503                     9.71                                             0.60
     twh_cons-solar_electricity twh_cons-wind_electricity
7475                       0.00                      0.00
7503                      23.26                     48.61

print(edm_pivot_longer2(inpt_datf = cur_data,
                        col_vars = c(4:ncol(cur_data)),
                        col_vars_to = "type_energie"))

    individual country                         year
1  France_1995  France                         1995
2  France_1995  France                         1995
3  France_1995  France                         1995
4  France_1995  France                         1995
5  France_1995  France                         1995
6  France_1995  France                         1995
7  France_1995  France                         1995
8  France_1995  France                         1995
9  France_1995  France                         1995
10 France_2023  France                         2023
11 France_2023  France                         2023
12 France_2023  France                         2023
13 France_2023  France                         2023
14 France_2023  France                         2023
15 France_2023  France                         2023
16 France_2023  France                         2023
17 France_2023  France                         2023
18 France_2023  France                         2023
                              type_energie twh_cons
1                      biofuel_electricity     1.82
2                         coal_electricity    24.18
3                          gas_electricity     3.84
4                        hydro_electricity    71.33
5                      nuclear_electricity   377.23
6                          oil_electricity     10.5
7  other_renewable_exc_biofuel_electricity     0.51
8                        solar_electricity        0
9                         wind_electricity        0
10                     biofuel_electricity      9.5
11                        coal_electricity     2.16
12                         gas_electricity    31.43
13                       hydro_electricity    53.19
14                     nuclear_electricity   335.65
15                         oil_electricity     9.71
16 other_renewable_exc_biofuel_electricity      0.6
17                       solar_electricity    23.26
18                        wind_electricity    48.61

}
