% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{fixer_nest_v}
\alias{fixer_nest_v}
\title{fixer_nest_v}
\usage{
fixer_nest_v(cur_v, pttrn_v, wrk_v)
}
\arguments{
\item{cur_v}{is the input vector}

\item{pttrn_v}{is the vector containing all the patterns that may be contained in cur_v}

\item{wrk_v}{is a vector containing all the indexes of cur_v taken in count in the function}
}
\description{
Retur the elements of a vector "wrk_v" (1) that corresponds to the pattern of elements in another vector "cur_v" (2) according to another vector "pttrn_v" (3) that contains the patterof elements.
}
\examples{

print(fixer_nest_v(cur_v=c("oui", "non", "peut-etre", "oui", "non", "peut-etre"), 
             pttrn_v=c("oui", "non", "peut-etre"), 
                  wrk_v=c(1, 2, 3, 4, 5, 6)))

#[1] 1 2 3 4 5 6

print(fixer_nest_v(cur_v=c("oui", "non", "peut-etre", "oui", "non", "peut-etre"), 
                 pttrn_v=c("oui", "non"), 
                  wrk_v=c(1, 2, 3, 4, 5, 6)))

#[1]  1  2 NA  4  5 NA

}
