% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{how_normal}
\alias{how_normal}
\title{how_normal}
\usage{
how_normal(inpt_datf, normalised = TRUE, mean = 0, sd = 1)
}
\arguments{
\item{inpt_datf}{is the input dataframe containing all the values in the first column and their frequency (normalised or no), in the second column}

\item{normalised}{is a boolean, takes TRUE if the frequency for each value is divided by n, FALSE if not}

\item{mean}{is the mean of the normal distribution that the dataset tries to fit}

\item{sd}{is the standard deviation of the normal distribution the dataset tries to fit}
}
\description{
Allow to get how much a sequence of numbers fit a normal distribution with chosen parameters, see examples
}
\examples{

sample_val <- round(rnorm(n = 12000, mean = 6, sd = 1.25), 1)
sample_freq <- unique_total(sample_val)
datf_test <- data.frame(unique(sample_val), sample_freq)
print(datf_test)

  unique.sample_val. sample_freq
1                 5.2         315
2                 4.6         208
3                 5.5         359
4                 6.8         307
5                 5.6         406
6                 7.2         240
7                 5.3         327
8                 6.0         364
9                 8.7          45
10                6.5         337
11                5.4         311
12                6.2         383
13                7.7         171
14                6.1         382
15                5.7         393
16                7.5         200
17                3.8          88
18                5.0         277
19                7.9          91
20                6.4         380
21                4.0          95
22                7.0         247
23                6.9         284
24                7.3         242
25                5.8         376
26                6.6         349
27                3.6          64
28                4.5         181
29                8.0          99
30                7.4         192
31                6.3         382
32                4.3         174
33                5.1         306
34                9.3          12
35               10.0           2
36                4.4         189
37                5.9         395
38                8.1          81
39                4.9         248
40                8.2          85
41                4.7         231
42                4.8         232
43                4.2         112
44                6.7         321
45                3.2          38
46                7.1         277
47                4.1         108
48                8.4          66
49                8.9          17
50                8.5          53
51                2.3           5
52                2.5           6
53                3.3          38
54                3.4          40
55                2.6           8
56                3.5          52
57                7.6         178
58                3.7          58
59                2.9          17
60                8.6          52
61                8.3          74
62                3.9          93
63                7.8         135
64               10.4           1
65                3.0          20
66               10.1           3
67                8.8          26
68                9.7           4
69                3.1          18
70                9.1          18
71                2.4           9
72                1.3           2
73                9.0          21
74                2.1           3
75                2.7          13
76                9.4           8
77                9.2           9
78                2.8          15
79                9.8           3
80                9.9           3
81                9.6          12
82                2.2           3
83                2.0           3
84               10.2           1
85                1.7           1
86                9.5           3
87                1.9           2
88               11.1           1

print(how_normal(inpt_datf = datf_test, 
                 normalised = FALSE,
                 mean = 6,
                 sd = 1))

[1] 9.002572

print(how_normal(inpt_datf = datf_test, 
                 normalised = FALSE,
                 mean = 5,
                 sd = 1))

[1] 9.098959

}
