% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{how_normal}
\alias{how_normal}
\title{how_normal}
\usage{
how_normal(inpt_datf, normalised = TRUE, mean = 0, sd = 1)
}
\arguments{
\item{inpt_datf}{is the input dataframe containing all the values in the first column and their frequency (normalised or no), in the second column}

\item{normalised}{is a boolean, takes TRUE if the frequency for each value is divided by n, FALSE if not}

\item{mean}{is the mean of the normal distribution that the dataset tries to fit}

\item{sd}{is the standard deviation of the normal distribution the dataset tries to fit}
}
\description{
Allow to get how much a sequence of numbers fit a normal distribution with chosen parameters, see examples
}
\examples{

sample_val <- round(rnorm(n = 12000, mean = 6, sd = 1.25), 1)
sample_freq <- unique_total(sample_val)
datf_test <- data.frame(unique(sample_val), sample_freq)
print(datf_test)

  unique.sample_val. sample_freq
1                 6.9         306
2                 8.3          63
3                 7.7         148
4                 5.6         363
5                 6.5         349
6                 4.6         202
7                 6.6         324
8                 6.7         335
9                 6.0         406
10                5.7         365
11                7.9         109
12                6.2         420
13                5.9         386
14                4.5         185
15                5.1         326
16                6.1         360
17                5.5         346
18                6.3         375
19                7.4         207
20                7.6         162
21                4.2         129
22                3.9         102
23                5.2         325
24                2.3           7
25                5.8         387
26                6.4         319
27                9.1          21
28                7.0         280
29                8.8          27
30                4.9         218
31                8.1          98
32                3.0          25
33                8.4          66
34                4.3         160
35                7.2         267
36                8.7          40
37                5.3         313
38                4.1         127
39                5.0         275
40                4.0         119
41                9.3          13
42                4.4         196
43                6.8         313
44                7.1         247
45                3.5          57
46                7.8         139
47                3.6          57
48                7.5         189
49                7.3         215
50                4.7         230
51                3.2          36
52                9.5           8
53                3.8          79
54                8.2          62
55                5.4         343
56                8.5          55
57                4.8         207
58                3.7          79
59                8.6          33
60                3.3          38
61                3.4          43
62                8.9          21
63                8.0         105
64                3.1          23
65                9.0          27
66               10.0           5
67                2.5          10
68                2.9          16
69                9.7           7
70                2.7          11
71               10.5           1
72                9.4          13
73                9.2          16
74                2.6          16
75                9.9           3
76                2.8          10
77                2.4          10
78                1.9           2
79                2.0           6
80               10.2           2
81                9.6           3
82               11.3           1
83                1.8           1
84                2.2           3
85                2.1           2
86                1.6           1
87               10.6           1
88                9.8           1
89               10.4           1
90                1.7           1

print(how_normal(inpt_datf = datf_test, 
                 normalised = FALSE,
                 mean = 6,
                 sd = 1))

[1] 9.003683

print(how_normal(inpt_datf = datf_test, 
                 normalised = FALSE,
                 mean = 5,
                 sd = 1))

[1] 9.098484

}
