% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{insert_df}
\alias{insert_df}
\title{insert_df}
\usage{
insert_df(df_in, df_ins, ins_loc)

insert_df(df_in, df_ins, ins_loc)
}
\arguments{
\item{df_in}{is the dataframe that will be inserted}

\item{df_ins}{is the dataset to be inserted
ins_loc is a vector containg two parameters (row, column) of the begining for the insertion}
}
\description{
Allow to insert dataframe into another dataframe according to coordinates (row, column) from the dataframe that will be inserted

Allow to insert dataframe into another dataframe according to coordinates (row, column) from the dataframe that will be inserted
}
