% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{just_nb}
\alias{just_nb}
\title{just_nb}
\usage{
just_nb(inpt_v, track_ = TRUE, symbol_ = "-")
}
\arguments{
\item{inpt_v}{is the input vector}

\item{track_}{is a boolean, if set to TRUE, the missing numbers are replaced by a symbol chosen}

\item{symbol_}{is the chosen symbol to replace numbers}
}
\description{
Extract only the letters from all elements of a vector, see examples
}
\examples{

print(just_nb(inpt_v = c("oui222jj644", "oui122jj"), 
    track_ = TRUE, symbol_ = "-"))

[1] "-222-44" "-122-"  

print(just_nb(inpt_v = c("oui222jj644", "oui122jj"), 
track_ = FALSE, symbol_ = "-"))

[1] "222644" "122"

}
