% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{letter_to_nb}
\alias{letter_to_nb}
\title{letter_to_nb}
\usage{
letter_to_nb(letter)
}
\arguments{
\item{letter}{is the letter (name of the column)}
}
\description{
Allow to get the number of a spreadsheet based column by the letter ex: AAA = 703
}
\examples{

print(letter_to_nb("rty"))

#[1] 12713

}
