% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{nb_to_letter}
\alias{nb_to_letter}
\title{nb_to_letter}
\usage{
nb_to_letter(x)

nb_to_letter(x)
}
\arguments{
\item{x}{is the number of the column}
}
\description{
Allow to get the letter of a spreadsheet based column by the number ex: 703 = AAA

Allow to get the letter of a spreadsheet based column by the number ex: 703 = AAA
}
