% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{pattern_generator}
\alias{pattern_generator}
\title{pattern_generator}
\usage{
pattern_generator(base_, from_, nb, hmn = 1, after = 1, sep = "")
}
\arguments{
\item{base_}{is the pattern that will be kept}

\item{from_}{is the vector from which the elements of the random part will be generated}

\item{nb}{is the number of random pattern chosen for the varying part}

\item{hmn}{is how many of varying pattern from the same base will be created}

\item{after}{is set to 1 by default, it means that the varying part will be after the fixed part, set to 0 if you want the varying part to be before}

\item{sep}{is the separator between all patterns in the returned value}
}
\description{
Allow to create patterns which have a part that is varying randomly each time.
}
\examples{

print(pattern_generator(base_="oui", from_=c("er", "re", "ere"), nb=1, hmn=3))

# [1] "ouier" "ouire" "ouier"

print(pattern_generator(base_="oui", from_=c("er", "re", "ere"), nb=2, hmn=3, after=0, sep="-"))

# [1] "er-re-o-u-i"  "ere-re-o-u-i" "ere-er-o-u-i"

}
