% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{see_df}
\alias{see_df}
\title{see_df}
\usage{
see_df(df, condition_l, val_l)

see_df(df, condition_l, val_l)
}
\arguments{
\item{df}{is the input dataframe}

\item{condition_l}{is the vector of the possible conditions ("==", ">", "<", "!=", "\%")}

\item{val_l}{is a list with the vector of values related to condition_l (so the values has to be placed in the same order)}

\item{conjunction_l}{contains the | or & conjunctions, so if the length of condition_l is equal to 3, there will be 2 conjunctions. If the length of conjunction_l is inferior to the length of condition_l minus 1, conjunction_l will match its goal length value with its last argument as the last arguments. For example, c("&", "|", "&") with a goal length value of 5 --> c("&", "|", "&", "&", "&")}
}
\description{
Allow to return a datafame with TRUE cells where the condition entered are respected and FALSE where these are not

Allow to return a datafame with TRUE cells where the condition entered are respected and FALSE where these are not
}
\examples{
see_df(df, c("\%\%", "=="), list(c(2, 11), c(3)), list("|") will return all the values that are divisible by 2 and 11 and all the values that are equal to 3 from the dataframe  
}
