% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{see_df}
\alias{see_df}
\title{see_df}
\usage{
see_df(df, condition_l, val_l)
}
\arguments{
\item{df}{is the input dataframe}

\item{condition_l}{is the vector of the possible conditions ("==", ">", "<", "!=", "\%")}

\item{val_l}{is the vector with the values related to condition_l (so the values has to be placed in the same order)}
}
\description{
Allow to return a datafame with TRUE cells where the condition entered are respected and FALSE where these are not
}
