% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{see_idx}
\alias{see_idx}
\title{see_idx}
\usage{
see_idx(v1, v2, exclude_val = "######", no_more = F)
}
\arguments{
\item{v1}{is the first vector}

\item{v2}{is the second vector}
}
\description{
Allow to find the indexes of the elements of the first vector in the second.
}
