% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{see_idx}
\alias{see_idx}
\title{see_idx}
\usage{
see_idx(v1, v2)
}
\arguments{
\item{v1}{is the first vector}

\item{v2}{is the second vector}
}
\description{
Returns a boolean vector to see if a set of elements contained in v1 is also contained in another vector (v2)
}
\examples{

print(see_idx(v1=c("oui", "non", "peut", "oo"), v2=c("oui", "peut", "oui")))

#[1]  TRUE FALSE  TRUE  FALSE

}
