% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{sort_normal_qual2}
\alias{sort_normal_qual2}
\title{sort_normal_qual2}
\usage{
sort_normal_qual2(inpt_datf)
}
\arguments{
\item{inpt_datf}{is the input dataframe, containing the values in the first column and their frequency in the second}
}
\description{
Sort qualitative modalities that have their frequency normally distributed from an unordered dataset, see examples. This function uses an another algorythm than choose_normal_qual which may be faster.
}
\examples{

sample_val <- round(rnorm(n = 2000, mean = 12, sd = 2), 1)
sample_freq <- unique_total(sample_val)
sample_qual <- infinite_char_seq(n = length(sample_freq))
datf_test <- data.frame(sample_qual, sample_freq)
datf_test[, 2] <- datf_test[, 2] / sum(datf_test[, 2])

print(datf_test)

   sample_qual sample_freq
1             a 0.208695652
2             b 0.234782609
3             c 0.321739130
4             d 0.339130435
5             e 0.330434783
6             f 0.069565217
7             g 0.234782609
8             h 0.400000000
9             i 0.347826087
10            j 0.043478261
11            k 0.278260870
12            l 0.286956522
13            m 0.243478261
14            n 0.147826087
15            o 0.234782609
16            p 0.252173913
17            q 0.417391304
18            r 0.095652174
19            s 0.313043478
20            t 0.008695652
21            u 0.130434783
22            v 0.391304348
23            w 0.113043478
24            x 0.295652174
25            y 0.243478261
26            z 0.382608696
27           aa 0.008695652
28           ab 0.347826087
29           ac 0.330434783
30           ad 0.321739130
31           ae 0.347826087
32           af 0.321739130
33           ag 0.173913043
34           ah 0.278260870
35           ai 0.278260870
36           aj 0.347826087
37           ak 0.026086957
38           al 0.295652174
39           am 0.226086957
40           an 0.295652174
41           ao 0.234782609
42           ap 0.113043478
43           aq 0.234782609
44           ar 0.173913043
45           as 0.017391304
46           at 0.252173913
47           au 0.078260870
48           av 0.086956522
49           aw 0.278260870
50           ax 0.086956522
51           ay 0.200000000
52           az 0.295652174
53           ba 0.052173913
54           bb 0.165217391
55           bc 0.408695652
56           bd 0.269565217
57           be 0.104347826
58           bf 0.391304348
59           bg 0.104347826
60           bh 0.043478261
61           bi 0.200000000
62           bj 0.095652174
63           bk 0.191304348
64           bl 0.008695652
65           bm 0.165217391
66           bn 0.226086957
67           bo 0.086956522
68           bp 0.017391304
69           bq 0.121739130
70           br 0.234782609
71           bs 0.121739130
72           bt 0.078260870
73           bu 0.173913043
74           bv 0.104347826
75           bw 0.208695652
76           bx 0.017391304
77           by 0.243478261
78           bz 0.034782609
79           ca 0.017391304
80           cb 0.008695652
81           cc 0.173913043
82           cd 0.147826087
83           ce 0.060869565
84           cf 0.017391304
85           cg 0.060869565
86           ch 0.008695652
87           ci 0.208695652
88           cj 0.043478261
89           ck 0.052173913
90           cl 0.017391304
91           cm 0.017391304
92           cn 0.095652174
93           co 0.113043478
94           cp 0.017391304
95           cq 0.017391304
96           cr 0.026086957
97           cs 0.034782609
98           ct 0.017391304
99           cu 0.026086957
100          cv 0.026086957
101          cw 0.026086957
102          cx 0.017391304
103          cy 0.043478261
104          cz 0.008695652
105          da 0.034782609
106          db 0.017391304
107          dc 0.060869565
108          dd 0.008695652
109          de 0.008695652
110          df 0.017391304
111          dg 0.008695652
112          dh 0.008695652
113          di 0.017391304
114          dj 0.008695652
115          dk 0.008695652

print(sort_normal_qual2(inpt_datf = datf_test))

0.00869565217391304 0.00869565217391304 0.00869565217391304 0.00869565217391304 
              "aa"                "cb"                "cz"                "de" 
0.00869565217391304 0.00869565217391304  0.0173913043478261  0.0173913043478261 
              "dh"                "dk"                "bp"                "ca" 
0.0173913043478261  0.0173913043478261  0.0173913043478261  0.0173913043478261 
              "cl"                "cp"                "ct"                "db" 
0.0173913043478261  0.0260869565217391  0.0260869565217391  0.0347826086956522 
              "di"                "cr"                "cv"                "bz" 
0.0347826086956522  0.0434782608695652  0.0434782608695652  0.0521739130434783 
              "da"                "bh"                "cy"                "ck" 
0.0608695652173913  0.0695652173913043  0.0782608695652174  0.0869565217391304 
              "cg"                 "f"                "bt"                "ax" 
0.0956521739130435  0.0956521739130435   0.104347826086957    0.11304347826087 
               "r"                "cn"                "bg"                 "w" 
  0.11304347826087   0.121739130434783   0.147826086956522   0.165217391304348 
              "co"                "bs"                 "n"                "bb" 
 0.173913043478261   0.173913043478261   0.191304347826087                 0.2 
              "ag"                "bu"                "bk"                "bi" 
 0.208695652173913   0.226086956521739   0.234782608695652   0.234782608695652 
              "bw"                "am"                 "b"                 "o" 
 0.234782608695652   0.243478260869565   0.243478260869565   0.252173913043478 
              "aq"                 "m"                "by"                "at" 
 0.278260869565217   0.278260869565217    0.28695652173913   0.295652173913043 
               "k"                "ai"                 "l"                "al" 
 0.295652173913043   0.321739130434783   0.321739130434783   0.330434782608696 
              "az"                 "c"                "af"                "ac" 
 0.347826086956522   0.347826086956522   0.382608695652174   0.391304347826087 
               "i"                "ae"                 "z"                "bf" 
 0.408695652173913   0.417391304347826                 0.4   0.391304347826087 
              "bc"                 "q"                 "h"                 "v" 
 0.347826086956522   0.347826086956522   0.339130434782609   0.330434782608696 
              "aj"                "ab"                 "d"                 "e" 
 0.321739130434783    0.31304347826087   0.295652173913043   0.295652173913043 
              "ad"                 "s"                "an"                 "x" 
 0.278260869565217   0.278260869565217   0.269565217391304   0.252173913043478 
              "aw"                "ah"                "bd"                 "p" 
 0.243478260869565   0.234782608695652   0.234782608695652   0.234782608695652 
               "y"                "br"                "ao"                 "g" 
 0.226086956521739   0.208695652173913   0.208695652173913                 0.2 
              "bn"                "ci"                 "a"                "ay" 
 0.173913043478261   0.173913043478261   0.165217391304348   0.147826086956522 
              "cc"                "ar"                "bm"                "cd" 
 0.130434782608696   0.121739130434783    0.11304347826087   0.104347826086957 
               "u"                "bq"                "ap"                "bv" 
 0.104347826086957  0.0956521739130435  0.0869565217391304  0.0869565217391304 
              "be"                "bj"                "bo"                "av" 
0.0782608695652174  0.0608695652173913  0.0608695652173913  0.0521739130434783 
              "au"                "dc"                "ce"                "ba" 
0.0434782608695652  0.0434782608695652  0.0347826086956522  0.0260869565217391 
              "cj"                 "j"                "cs"                "cw" 
0.0260869565217391  0.0260869565217391  0.0173913043478261  0.0173913043478261 
              "cu"                "ak"                "df"                "cx" 
0.0173913043478261  0.0173913043478261  0.0173913043478261  0.0173913043478261 
              "cq"                "cm"                "cf"                "bx" 
0.0173913043478261 0.00869565217391304 0.00869565217391304 0.00869565217391304 
              "as"                "dj"                "dg"                "dd" 
0.00869565217391304 0.00869565217391304 0.00869565217391304 
              "ch"                "bl"                 "t" 

}
