% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{unique_pos}
\alias{unique_pos}
\title{unique_pos}
\usage{
unique_pos(vec)

unique_pos(vec)
}
\arguments{
\item{vec}{is the input vector}
}
\description{
Return the indexes of the first unique values from a vector

Return the indexes of the first unique values from a vector
}
