% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{unique_pos}
\alias{unique_pos}
\title{unique_pos
Allow to find indexes of the unique values from a vector.}
\usage{
unique_pos(vec)
}
\arguments{
\item{vec}{is the input vector}
}
\description{
unique_pos
Allow to find indexes of the unique values from a vector.
}
