% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R, R/all_fun_quarantine.R
\name{val_replacer}
\alias{val_replacer}
\title{val_replacer}
\usage{
val_replacer(df, val_replaced = F, val_replacor = T, df_rpt = NA)

val_replacer(df, val_replaced = F, val_replacor = T, df_rpt = NA)
}
\arguments{
\item{df}{is the input dataframe}

\item{val_replaced}{is a vector of the value(s) to be replaced}

\item{val_replacor}{is the value that will replace val_replaced}

\item{df_rpt}{is the replacement matrix and has to be the same dimension as df. Only the indexes that are equal to TRUE will be authorized indexes for the values to be replaced in the input matrix}
}
\description{
Allow to replace value from dataframe to another one.

Allow to replace value from dataframe to another one.
}
