% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{val_replacer}
\alias{val_replacer}
\title{val_replacer}
\usage{
val_replacer(datf, val_replaced, val_replacor = TRUE)
}
\arguments{
\item{datf}{is the input dataframe}

\item{val_replaced}{is a vector of the value(s) to be replaced}

\item{val_replacor}{is the value that will replace val_replaced}
}
\description{
Allow to replace value from dataframe to another one.
}
\examples{

print(val_replacer(datf=data.frame(c(1, "oo4", TRUE, FALSE), c(TRUE, FALSE, TRUE, TRUE)), 
     val_replaced=c(TRUE), val_replacor="NA"))

#  c.1...oo4...T..F. c.T..F..T..T.
#1                 1            NA
#2               oo4         FALSE
#3                NA            NA
#4             FALSE            NA

}
