% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{vec_in_df}
\alias{vec_in_df}
\title{vec_in_df}
\usage{
vec_in_df(df_, vec_l, coeff_, strt_l, distinct = "NA")
}
\arguments{
\item{df_}{is the input dataframe}

\item{vec_l}{is a list the vectors}

\item{coeff_}{is the related coefficient of the vector}

\item{strt_l}{is a vector containing the start position for each vector}

\item{distinct}{is a value you are sure is not in df_, defaults to "NA"}
}
\description{
Allow to see if vectors are present in a dataframe
ex: 1, 2, 1
3, 4, 1
1, 5, 8
the vector c(4, 1) with the coefficient 1 and the start position at the second column is containded in the dataframe
}
