% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_fun.R
\name{vlookup_df}
\alias{vlookup_df}
\title{vlookup_df}
\usage{
vlookup_df(df, v_id, col_id = 1, included_col_id = "yes")
}
\arguments{
\item{df}{is the input dataframe}

\item{v_id}{is a vector containing the ids}

\item{col_id}{is the column that contains the ids (default is equal to 1)}

\item{included_col_id}{is if the result should return the col_id (default set to yes)}
}
\description{
Alow to perform a vlookup on a dataframe
}
\examples{
df1 <- data.frame(c("az1", "az3", "az4", "az2"), c(1:4), c(4:1))

print(vlookup_df(df=df1, v_id=c("az1", "az2", "az3", "az4")))

   c..az1....az3....az4....az2.. c.1.4. c.4.1.
2                            az1      1      4
4                            az2      4      1
21                           az3      2      3
3                            az4      3      2
}
